% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{adjustIncidence}
\alias{adjustIncidence}
\title{Legacy nowcasting interface}
\usage{
adjustIncidence(
  obj,
  datas = NULL,
  method = "none",
  pdig = stats::plnorm((1:20) * 7, 2.5016, 1.1013),
  Dmax = 10,
  nyears = 2,
  datasource = NULL,
  nowSE,
  safelimit = 5,
  seed = NULL,
  workers = 1L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{obj, datas, method, pdig, Dmax, nyears, datasource, nowSE, safelimit, seed, workers, verbose, ...}{Legacy arguments.}
}
\description{
Legacy nowcasting interface
}
