% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{alertaRio}
\alias{alertaRio}
\title{4 level alert Green-Yellow-Orange-Red for Rio de Janeiro.}
\usage{
alertaRio(naps = 0:9, pars, crit, datasource, se, verbose = TRUE)
}
\arguments{
\item{naps}{subset of vector 0:9 corresponding to the id of the APS. Default is all of them.}

\item{pars}{parameters of the alert.}

\item{datasource}{it is the name of the sql connection.}

\item{se}{last epidemiological week (format = 201401)}

\item{verbose}{FALSE}
}
\value{
list with an alert object for each APS.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
params <- list(pdig = c(2.5016,1.1013), tcrit=22, inccrit=100, preseas = 14.15, posseas = 18)
criter <- criteria = list(
crity = c("temp_min > tcrit | (temp_min < tcrit & inc > preseas)", 3, 0),
crito = c("p1 > 0.9 & inc > preseas", 2, 2),
critr = c("inc > inccrit", 1, 2)
)
alerio2 <- alertaRio(naps = c(0,1), pars=params, crit = criter, datasource=con, se=201604)
names(alerio2)
}

