\name{AlertTools-package}
\alias{AlertTools-package}
\alias{AlertTools}
\docType{package}
\title{
\packageTitle{AlertTools}
}
\description{
\packageDescription{AlertTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlertTools}
\packageIndices{AlertTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AlertTools}

Maintainer: \packageMaintainer{AlertTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
