% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlertTools.R
\docType{package}
\name{AlertTools}
\alias{AlertTools}
\alias{AlertTools-package}
\title{AlertTools: utilities for the InfoDengue alert system}
\description{
AlertTools retrieves and organizes epidemiological and climate data,
adjusts notification series, estimates epidemiological indicators, and
calculates alert levels. Database access is isolated behind an internal DBI
repository, while computational functions operate on in-memory data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlertaDengue/AlertTools}
  \item Report bugs at \url{https://github.com/AlertaDengue/AlertTools/issues}
}

}
\author{
\strong{Maintainer}: Claudia Codeco \email{claudia.codeco@gmail.com}

}
\keyword{internal}
