% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue-mem.R
\name{applymem}
\alias{applymem}
\title{Get epidemic thresholds  for a list of cities}
\usage{
applymem(df.data, l.seasons, typicalcurve = FALSE, ...)
}
\arguments{
\item{df.data}{data frame with municipio_geocodigo in a column named municipio_geocodigo,
and incidence seasons in each column each row gives the incidence in each season,
 for each municipio_geocodigo, for each week.}

\item{l.seasons}{vector with incidence columns to be used. Additional arguments
accepted by memmodel function}

\item{typicalcurve}{Whether to retain the typical epidemic curve.}

\item{...}{Additional arguments passed to `mem::memmodel()`.}
}
\value{
epithresholds: list with full epimem report for each municipio_geocodigo,
keyed by AP's name. dfthresholds: data frame with thresholds for each AP.
}
\description{
Function to apply epimem algorithm on df.data and generate full
reports as well as a data frame with summary of relevant thresholds (pre, pos, mid, high, veryhigh)
}
\details{
internal function, used by info.dengue.apply.mem.   Defatul arguments
passed to memmodel function:
}
\author{
Marcelo F Gomes
}
