% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{as_epiweek}
\alias{as_epiweek}
\title{Convert dates to epidemiological weeks}
\usage{
as_epiweek(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{A vector coercible to `Date`.}

\item{format}{Format used for character input.}
}
\value{
An integer vector in `YYYYWW` format.
}
\description{
Uses the Brazilian Sunday-to-Saturday convention in which epidemiological
week 1 is the week containing 4 January.
}
\examples{
as_epiweek(as.Date(c("2020-12-27", "2021-01-03")))
}
