% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{bestWU}
\alias{bestWU}
\title{Chooses the most updated meteorological data for a city from a set of two choices}
\usage{
bestWU(series, var)
}
\arguments{
\item{series}{list of the competing time series.
See examples.}

\item{var}{name of the meteorological variable}
}
\value{
data.frame with the best series
}
\description{
Receives a set of meteorological data and return the most complete for
a given city.If both are good, the first is returned.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# NOT USE: connection <- dbConnect(RSQLite::SQLite(), "../../AlertaDengueAnalise/mydengue.sqlite")
series1 = getWU(stations = 'SBCP',vars="temp_min", datasource= connection)
series2 = getWU(stations = 'SBME',vars="temp_min", datasource= connection)
res = bestWU(list(series1, series2),var="temp_min")
tail(res)
\dontshow{\}) # examplesIf}
}
