% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{calc.alerta}
\alias{calc.alerta}
\title{Calculates the attention levels for one area.}
\usage{
calc.alerta(x, pars, level = "municipio", ...)
}
\arguments{
\item{x}{filter variable indicating geocode at municipal, state or other level.}

\item{pars}{List containing input data, report week, nowcasting mode, parameters, and case records.}

\item{level}{default is municipio. Other options are: regional, macrorregiao
and uf (not implemented yet)}

\item{...}{Reserved for compatibility.}
}
\value{
data.frame with the data with alert level.
}
\description{
Function to compute the alert levels for one spatial unit, defined by level.
Internal function used by the dengue_pipeline.
}
