% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{calc_indicators}
\alias{calc_indicators}
\title{Compute indicators}
\usage{
calc_indicators(ale)
}
\arguments{
\item{ale}{alert object from one city}
}
\value{
vector of receptivity
}
\description{
Calculate climate receptivity
}
\examples{
res <- pipe_infodengue(cities = 3304557, cid10 = "A90", 
finalday= "2020-01-23",nowcasting="none")
calc_indicators(res)
}
