% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{casesinlocality}
\alias{casesinlocality}
\title{Get disagreggated time series of cases and aggregate per health district}
\usage{
casesinlocality(obj, locality)
}
\arguments{
\item{obj}{data.frame created by getCases with withdivision = TRUE (per bairro)}

\item{division}{data.frame with bairros and corresponding health districts. The neighborhood
names must match exactly the ones in the sinan. If "test" uses test data from Rio.}
}
\value{
data.frame with the data aggregated per health district and week
}
\description{
Get time series of cases generated by getCases withdivision=TRUE 
and aggregate it per health district. Requires a file containing the list of 
neighborhoods in the city and their corresponding health districts.
}
\examples{
dC0 = getCases(city = c(330455), withdivision=TRUE,datasource = "data/sinan.rda") # Rio de Janeiro
dC1 = casesinlocality(obj = dC0, locality = "AP1") 
Gives an error message: dataframe contains no column BAIRRO. 
dC0 = getCases(city = c(330455), withdivision = FALSE,datasource = "data/sinan.rda") 
dC1 = casesinlocality(obj = dC0, locality = "AP1") # Rio de Janeiro
}

