% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{classify}
\alias{classify}
\title{Classifies the disease time series into transmission levels.}
\usage{
classify(yel, ora = c(), red = c(), inerciaon = 3, inerciaoff = c(NA, 3,
  3, 3))
}
\arguments{
\item{red}{object from setred function.}

\item{yellow}{object from setyellow function.}

\item{orange}{object from setorange function.}
}
\value{
data.frame with
}
\description{
This function receives the time series inputs (cases, tweets,
climate) and return levels of alert (1 (low) to 4 (high)). It will run the
set color functions and them combine them hierarchically.
}
\examples{
#This is the whole sequence of the alert. Getting data:
res <- getCases(city = c(330455), withdivision = TRUE) # get case data from Rio de Janeiro
resd <- aggrbylocality(d = res, locality="AP1") # case data from a locality in Rio de Janeiro
tweet <- getTweet(city = c(330455)) # Rio de Janeiro # get tweeter data
clima <- getWU(stations = 'SBRJ', var = "tmin") # get locality's tmin from local meteorological station
#Adjust incidence
adjcase <- adjustIncidence(se = resd$SE, y = resd$casos)
# Calculate Rt
rtnorm<-Rt(obj = resfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
#Calculate each level
yellowalert <- setYellow(temp = clima$tmin, se = clima$SE, tempcrit = 22)
orangealert <- setOrange(obj = rtnorm, pvalue = 0.9, lag= 3)
redalert <- setRed(obj = resfit, pop = 30000, ccrit = 100, lag=3)
# green - yellow levels only
resu <- classify(yel = yellowalert) # with only green-yellow levels
tail(resu)
x = 1:length(resu$temp)
plot(x,resu$temp, type="l", xlab= "weeks", ylab = "temp")
points(x[resu$level==2], resu$temp[resu$level==2], col="yellow",pch=16)
# all four levels
resu <- classify(yel = yellowalert, ora = orangealert, red = redalert, inerciaon=1, inerciaoff=c(NA,1,1,1))
x = 1:length(resu$casos)
plot(x,resu$casos, type="l", xlab= "weeks", ylab = "casos")
points(x[resu$level==2], resu$casos[resu$level==2], col="yellow",pch=16)
points(x[resu$level==3], resu$casos[resu$level==3], col="orange",pch=16)
points(x[resu$level==4], resu$casos[resu$level==4], col="red",pch=16)
}

