% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{count_transmission_weeks}
\alias{count_transmission_weeks}
\title{Counts the number of successive weeks with Rt > 1}
\usage{
count_transmission_weeks(d, alpha = 0.1, maxw = 52)
}
\arguments{
\item{d}{dataset with data to feed the alert. It requires the previous 
computation of Rt.}

\item{alpha}{significance level for considering Rt > 1 (default = 0.1)}

\item{maxw}{maximum value for the indicator (default = 30)}
}
\value{
returns the data, and the number of weeks with transmission.
}
\description{
Indicator used for raising the orange alert
}
\examples{
# Parameters of the alert model (requires connection)
casos <- getCases(c(3302205, 3304557), cid10 = "A90", type = "all", 
completetail = 0, dataini = "sinpri") 
casos <- casos \%>\% Rt(count = "casos", group = "cidade", 
gtdist="normal", meangt=3, sdgt = 1) 
casos <- casos \%>\% count_transmission_weeks()
}
