% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd}
\alias{dd}
\title{dengue notification in Acre}
\format{
## `dd`
A data frame with 12,188 rows and 30 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\usage{
dd
}
\description{
A subset of the SINAN data , to be used as example
}
\keyword{datasets}
