% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{delaycalc}
\alias{delaycalc}
\title{Organize delay data for analysis and produce a nice plot.}
\usage{
delaycalc(d, tini = "dt_notific", tfim = "dt_digita", SE = "SE_notif",
  date.format = "\%Y-\%m-\%d", truncdays = 183, verbose = TRUE)
}
\arguments{
\item{d}{dataset with case data containing at least three variables: the initial and final dates and a variable
identifying the epidemiological week. Only one city at a time.}

\item{tini}{variable indicating the initial date}

\item{tfim}{variable indicating the end date}

\item{SE}{variable indicating the epidemiological week}

\item{date.format}{date format. Default is "%d-%m-%Y"}

\item{truncdays}{Default is 183 days}

\item{plotar}{Default is TRUE}
}
\value{
list with d = data.frame with the epidemiological weeks; delay.tbl and delay.week 
are internal objects used for plotting.
}
\description{
The second function to be used in the delay fitting process using inla. Calculates the number
of cases reported per week with a given delay. Also removes data with notification delay greater than truncdays.
}
\examples{
dados <- getdelaydata(cities=3302205, datasource=con)
res = delaycalc(dados)  
head(res$d)  # data
head(res$delay.tbl)  # running matrix
}
\author{
Leo Bastos
}
