% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{DenguedbConnect}
\alias{DenguedbConnect}
\title{Returns the connection to the database.}
\usage{
DenguedbConnect()
}
\value{
"PostgreSQLConnection" object
}
\description{
Opens a connection to the Project database.
}
\examples{
con <- DenguedbConnect()
dbListTables(con) 
dbDisconnect(con)
}

