% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek_seq}
\alias{epiweek_seq}
\title{Create an inclusive epidemiological-week sequence}
\usage{
epiweek_seq(from, to)
}
\arguments{
\item{from}{First epidemiological week in `YYYYWW` format.}

\item{to}{Last epidemiological week in `YYYYWW` format.}
}
\value{
An integer vector of epidemiological weeks.
}
\description{
Create an inclusive epidemiological-week sequence
}
\examples{
epiweek_seq(202052, 202101)
}
