% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek_start}
\alias{epiweek_start}
\title{Return the start date of epidemiological weeks}
\usage{
epiweek_start(x)
}
\arguments{
\item{x}{Epidemiological weeks in `YYYYWW` format.}
}
\value{
A `Date` vector containing Sundays.
}
\description{
Return the start date of epidemiological weeks
}
\examples{
epiweek_start(c(202052, 202101))
}
