% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{epiYear}
\alias{epiYear}
\title{Define Epidemiological Year.}
\usage{
epiYear(se, cut = 41)
}
\arguments{
\item{se}{numeric vector with epidemiological weeks to be converted}

\item{cut}{epidemiological week that separates consecutive epidemiological years. Default = 41}
}
\value{
vector of epidemiological years.
}
\description{
Find to which epidemiological year belongs a given epidemiological week.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
epiYear(se = 201012)
epiYear(se = 201012:201522)
\dontshow{\}) # examplesIf}
}
