% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{fitDelay.inla}
\alias{fitDelay.inla}
\title{Fit INLA model to notification delay.}
\usage{
fitDelay.inla(obj, Tactual = nrow(obj$delay.tbl), Dmax = 12,
  plotar = FALSE)
}
\arguments{
\item{obj}{object with observed delays, produced by delaycalc().}

\item{Tactual}{today's date used for estimation.}

\item{Dmax}{maximum delay allowed (in weeks).}
}
\value{
list containing the fitted model (out), the data, arguments and some objects to be used
by other auxiliary functions.
}
\description{
Fit model  Y ~ 1 + f(Time, model = "rw1") + f(Delay, model = "rw1") to object created by delaycalc.
}
\examples{
res <- delaycalc(dados)
outp<-fitDelay.inla(res)
}
