% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fouralert}
\alias{fouralert}
\title{Legacy alert classifier}
\usage{
fouralert(obj, crit, miss = "last", dy = 4)
}
\arguments{
\item{obj, crit, miss, dy}{Legacy arguments.}
}
\description{
Legacy alert classifier
}
