% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{geraMapa}
\alias{geraMapa}
\title{Plot the alert map for any state.}
\usage{
geraMapa(alerta, subset, cores = c("green", "yellow", "orange", "red"),
  legpos = "bottomright", se, datasource = con, shapefile, varid, varname,
  titulo = "", filename, dir = "", caption = TRUE, resol = 200)
}
\arguments{
\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}

\item{legpos}{legend position. Default is bottomright. Check the other options in the function legend.}

\item{se}{epidemiological week (format = 201610).}

\item{shapefile}{shapefile containing polygons for the municipalities}

\item{varid}{shapefile variable indicating the geocode of the municipalities}

\item{varname}{name of the variable to be plotted}

\item{titulo}{title of the map}

\item{filename}{if present, the map is saved.}

\item{dir}{directory where map will be saved}

\item{resol}{dpi png resolution, default is 200}

\item{regionais}{vector of alerts created by update.alerta.}
}
\value{
a map
}
\description{
Function to plot a map of the alert
}
\examples{
reg1 <- update.alerta(region="Noroeste", pars = pars.RJ, datasource = con, sefinal=201613)
geraMapa(reg1, data=201613, shapefile="../report/RJ/shape/33MUE250GC_SIR.shp",
se = 201613, varid="CD_GEOCMU", titulo="Regionais Norte e Nordeste \\n", legpos="topright")
}

