% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getCases}
\alias{getCases}
\title{Legacy case reader}
\usage{
getCases(
  cities,
  lastday = Sys.Date(),
  firstday = as.Date("2018-01-01"),
  cid10 = "A90",
  dataini = "notific",
  completetail = NA,
  type = "notified",
  datasource,
  verbose = FALSE
)
}
\arguments{
\item{cities, lastday, firstday, cid10, dataini, completetail, type, datasource}{Legacy arguments.}

\item{verbose}{Whether to emit progress messages.}
}
\description{
Legacy case reader
}
