% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCasesinRio}
\alias{getCasesinRio}
\title{Get cases from an APS in Rio de Janeiro and aggregate them into weekly time series.}
\usage{
getCasesinRio(APSid, lastday = Sys.Date(), disease = "dengue", datasource)
}
\arguments{
\item{APSid}{0(APS1), 1 (APS2.1), 2 (APS2.2), 3(APS3.1), 4(APS3.2), 5(APS3.3), 6(APS4),
, 7(APS5.1), 8(APS5.2), 9(APS5.3)}
}
\value{
data.frame with the data aggregated per health district and week
}
\description{
Get time series of cases per APS in Rio de Janeiro (special case)
}
\examples{
dC = getCasesinRio(APSid = 1, datasource = con) # Rio de Janeiro
tail(dC)
}

