% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getCaseslist}
\alias{getCaseslist}
\title{Legacy individual-case reader}
\usage{
getCaseslist(
  cities,
  lastday = Sys.Date(),
  firstday = as.Date("2022-01-01"),
  cid10 = "A90",
  datasource
)
}
\arguments{
\item{cities, lastday, firstday, cid10, datasource}{Legacy arguments.}
}
\description{
Legacy individual-case reader
}
