% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{getdelaydata}
\alias{getdelaydata}
\title{Get delay data for one or more cities for delay analysis}
\usage{
getdelaydata(
  cities,
  nyears = 2,
  cid10 = "A90",
  lastday = Sys.Date(),
  datasource
)
}
\arguments{
\item{cities}{vector with geocodes}

\item{nyears}{number of years of data used for fitting: Default = 2 years}

\item{cid10}{disease code, Default is dengue. "A92.0" for chik, "A92.8" for zika}

\item{lastday}{last digitation day}

\item{datasource}{valid connection to database}
}
\value{
list with d = data.frame with data.
}
\description{
Gets delay data for one or more cities. Internal function used in
the delay fitting using inla.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
dados <- getdelaydata(cities=3304557, nyears=1, cid10="A90",
lastday = as.Date("2023-12-31"))  # Not run without connection
\dontshow{\}) # examplesIf}
}
