% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tentative-functions.R
\name{getmergeInfodata}
\alias{getmergeInfodata}
\title{Get and merge cases, tweets and climate data.}
\usage{
getmergeInfodata(city, cid10, seini, sefinal = Sys.Date(), datasource)
}
\arguments{
\item{cases}{data.frame with aggregated cases by locality (or city)
and epidemiological week.}

\item{tweet}{data.frame with tweets aggregated per week}

\item{climate}{data.frame with climate data aggregated per week for the
station of interest.}
}
\value{
data.frame with all data available
}
\description{
Get and merge cases, tweets and climate data for the alert
}
\examples{
cas = getCases(city = c(330455), withdivision = FALSE, datasource = "data/sinan.rda") 
tw = getTweet(city = c(330455), datasource = "data/tw.rda")
clima = getWU(stations = 'SBRJ', var="temp_min", datasource="data/WUdata.rda")
head(mergedata(cases = cas, tweet = tw, climate = clima))
head(mergedata(tweet = tw, climate = clima))
head(mergedata(cases = cas, climate = clima))
head(mergedata(tweet = tw, cases = cas))
}
