% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.dengue.apply.mem.R
\name{info.dengue.apply.mem}
\alias{info.dengue.apply.mem}
\title{Function to apply MEM algorithm to generate pre-epidemic, post-epidemic, and high activity thresholds
to municipalities covered by InfoDengue project}
\usage{
info.dengue.apply.mem(mun_list = mun_list, start_year = 0,
  end_year = as.integer(format(Sys.Date(), "\%Y")) - 1, write = "no",
  con = NULL, passwd = NULL, i.n.max = 0, limiar.preseason = 0.9,
  limiar.epidemico = 0.95, ...)
}
\arguments{
\item{mun_list}{Vector of municipalities geocode. Default: NULL (i.e., all municipalites)}

\item{start_year}{Historical data starting year to consider. Default: 0 (i.e., as old as possible)}

\item{end_year}{Historical data final year to consider. Default: previous year from Sys.Date()}

\item{write}{Whether results should be saved into project's database. Default: 'no'
'no' return data.table without writing in database.
'db' writes data.table to InfoDengue's database using \code{write.parameters} from AlerttTools package.}

\item{con}{Connection to PostGreSQL database, using \code{dbConnect} from RPostgreSQL package}

\item{passwd}{database password for writing output to db if write='db'.}

\item{i.n.max}{Number of points by seasons to be used for pre-epidemic and epidemic regions to calculate each threshold.
If 0 (default), uses all points in those regions. Else, uses n max values in each region per season.
This value is passed to i.n.max parameter in \code{memmodel}, from MEM package.}

\item{limiar.preseason}{Level of confidence interval to calculate pre/post-season thresholds. Default: 0.9.
This value is passed to i.level.threshold parameter in \code{memmodel}, from MEM package.}

\item{limiar.epidemico}{Level of confidence interval to calculate high activity during epidemic region. Default: 0.95.
This value is passed to i.level.intensity parameter in \code{memmodel}, from MEM package.}

\item{...}{Optional arguments passed to \code{memmodel}, from MEM package.}
}
\value{
Function \code{info.dengue.apply.mem} will return a data.table object with results by municipality:
\item{thresholds.table}{A data.table object with results for each municipality geocode, with columns
municipio_geocodigo, ano_inicio, ano_fim, limiar_preseason, limiar_posseason, limiar_epidemico}
}
\description{
Function \code{info.dengue.apply.mem} uses MEM package to generate activity thresholds.
Apply MEM algorithm discarding seasons below local pre-epidemic threshold.
For each Municipality, pre-epidemic (limiar_preseason) threshold has a minimum set at 5 cases.
If calculated value falls below that, it is set to 5, with high activity (limiar_epidemico) set to 10 cases.
}
\examples{
Generate thresholds for Rio de Janeiro, Curitiba and Vitoria, using the whole history.
Return object instead of writing to data base:
mun_list <- c(3304557, 4106902, 3205309)
mun_list <- c(3205309)

thresholds.table <- info.dengue.apply.mem(con=con, passwd=password, start_year=0, mun_list=mun_list)

Write to database instead of returning object:
info.dengue.apply.mem(con=con, passwd=password, start_year=0, mun_list=mun_list, output='db')
}
