% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{insert_city_infodengue}
\alias{insert_city_infodengue}
\title{Initial setup of a new city in the alerta system. Insert into tables Regionais.}
\usage{
insert_city_infodengue(
  geocodigo,
  id_regional,
  regional,
  macroreg,
  datasource = con
)
}
\arguments{
\item{id_regional}{numerical id of the 'Regional da saude'.Mandatory}

\item{regional}{name of the 'Regional da saude'.Mandatory}

\item{macroreg}{name of the 'MacroRegional da saude'.Mandatory}

\item{city}{geocode of the city. Mandatory}
}
\value{
to be defined
}
\description{
Initial setup of a new city in the alerta system.  Can be integrated later with 
the delay model and write.parameters
}
\examples{
insert_city_infodengue(geocodigo = 1111111, id_regional=1, regional = "teste", macroreg = "teste")
}
