% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isreceptive.R
\name{is.receptive}
\alias{is.receptive}
\title{Check if conditions are receptive for transmission.}
\usage{
is.receptive(rule = NULL, data = NULL)
}
\arguments{
\item{rule}{string with a condition, ex: 'temp_min_1 > 22 & inc_3 > 1.1 | temp_min_1 < 22 & umid_max_2 > 90'.
where temp_min, inc, umid_max are variables in the data, while _i at the end indicate the time lag.}

\item{data}{data.frame containing at least the variables mentioned in the rule,
without the lagging. Lagging works only for inc_3, temp_min, temp_max, umid_min,
umid_max, with 1 to 3 weeks of lagging.}
}
\value{
vector with three possible values: 1 - receptive, 2 - non_receptive, or NA if not enough information      
the data with the lagged variables and a summary of the data quality.
}
\description{
receptivity is defined as the combination of suspected viral 
activity and favorable climate.
}
\examples{
rul <- c("temp_min_1 > 22 & inc_3 > 0")
dat <- data.frame(temp_min = rnorm(100,25,5), umid_max = rnorm(100,70, 8), inc = rpois(100, 2)/1000)
rec <- is.receptive(rule = rul, data = dat)
Using infodengue rules:
val <- read.parameters(1200401)
setCriteria(rule=val$rule, data = data)
}
