% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{is_receptive}
\alias{is_receptive}
\title{Define conditions to issue Yellow alert.}
\usage{
is_receptive(d, crit = rule)
}
\arguments{
\item{d}{dataset with data to feed the alert, containing the variables specified in crit.
only one place at a time.}

\item{crit}{criteria for receptivity. 
Ex: "(inc_3 > 4 & temp_min_1 > 20) | (inc_3 > 0 & temp_min_2 > 30)". Valid variables 
are: inc, temp_min, temp_max, umid_min, umid_max, with 0 to 3 time lags.}

\item{values}{vector with the required variables}

\item{miss}{how missing data is treated. "last" if last value is repeated. 
It is currently the only option}
}
\value{
returns the data, and the receptivity (1 = receptive, 0 = not receptive).
}
\description{
Yellow is raised when environmental and epidemiological conditions
required for transmission are detected.
}
\examples{
# Parameters of the alert model
rule <- "(inc_3 > 4 & temp_min_1 > 20) | (inc_3 > 0 & temp_min_2 > 30)"
dat <- data.frame(SE = 201001:201052, inc = rnorm(52, 50, 20), temp_min = rnorm(52, 20, 5),
temp_max = rnorm(52, 25, 5), umid_min = rnorm(52, 60, 10), umid_max = rnorm(52, 80, 5))
dat <- is_receptive(dat, rule)
# Using Alertools data example:
data <- AlertTools::dd \%>\% filter(codigo_regional == 12002)
data <- lag_variables(data)
crit <- AlertTools::regras
data <- data \%>\% is_receptive(crit = crit$regra[1])
table(data$receptivo)
}
