% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{lag_variables}
\alias{lag_variables}
\title{compute time lagged variables}
\usage{
lag_variables(
  d,
  group = "cidade",
  vars = c("temp_min", "temp_max", "umid_min", "umid_max", "inc"),
  lags = 1:3
)
}
\arguments{
\item{d}{dataset with the variables to be lagged}

\item{group}{column containing the geocode or other geographical identifier. For now,
only works with "cidade".}

\item{vars}{variables to be lagged}

\item{lags}{lags. default: 1:3}
}
\value{

}
\description{
Compute time lagged variables for the receptive function
}
\examples{
summary(dd) # dd is a dataset with all variables used by Infodengue
ddl <- lag_variables(dd)
}
