% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{lastDBdate}
\alias{lastDBdate}
\title{Returns the most recent date present in the database table.}
\usage{
lastDBdate(tab, cities, cid10 = "A90", stations, datasource = con)
}
\arguments{
\item{tab}{table in the database. Either (sinan, clima_wu, tweet, historico,
historico_mrj).}

\item{cities}{vector of geocodes.}

\item{cid10}{relevant for sinan, tweeter or historico. Codes are: dengue "A90", 
chik "A92.0", zika "A92.8".}

\item{stations}{vector with wu stations. Ex. c("SGBL", "SBRL")}

\item{datasource}{}
}
\value{
vector with two elements: data.max = most recent date in the collection of cities or stations;
se = corresponding epidemiological week.
}
\description{
Useful to check if the database is up-to-date.
}
\examples{
cidades <- getCidades(regional = "Sete Lagoas", uf = "Minas Gerais")
lastDBdate(tab = "tweet", cities = cidades$municipio_geocodigo, cid10 = "A90")
lastDBdate(tab = "sinan", cities = cidades$municipio_geocodigo, cid10 = "A92.8")
lastDBdate(tab = "clima_wu", stations = "SBAF", datasource=con)  
lastDBdate(tab = "clima_wu", cities = cidades$municipio_geocodigo, datasource=con)  
}
