% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{lastepiweek}
\alias{lastepiweek}
\title{lastepiweek -----------------------------------}
\usage{
lastepiweek(ano)
}
\arguments{
\item{ano}{Year}
}
\description{
Calculate number of year's last epidemiological week using Brazilian standard.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
lastepiweek(2018)
\dontshow{\}) # examplesIf}
}
\author{
Marcelo F Gomes
}
\keyword{internal}
