% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{map.Rio}
\alias{map.Rio}
\title{Plot the alert map for Rio de Janeiro city.}
\usage{
map.Rio(obj, cores = c("green", "yellow", "orange", "red"), data,
  datasource = con,
  shapefile = "../report/Rio_de_Janeiro/shape/CAPS_SMS.shp", filename = "",
  dir, resol = 200)
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}

\item{filename}{if present, the map is saved.}

\item{dir}{directory where map will be saved.}

\item{resol}{dpi png resolution, default is 200}

\item{var}{to be ploted in the graph, usually cases when available.}
}
\value{
a map
}
\description{
Function to plot a map of the alert
}
\examples{
map.Rio(alerio)
}

