% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{mergedata}
\alias{mergedata}
\title{Merge cases, tweets and climate data.}
\usage{
mergedata(cases = c(), tweet = c(), climate = c())
}
\arguments{
\item{cases}{data.frame with aggregated cases by locality (or city)
and epidemiological week.}

\item{tweet}{data.frame with tweets aggregated per week}

\item{climate}{data.frame with climate data aggregated per week for the
station of interest.}
}
\value{
data.frame with all data available
}
\description{
Merge cases, tweets and climate data for the alert
}
\examples{
cas = getCases(city = c(330455), withdivision = FALSE)
casa = getCases(city = c(330455), withdivision = TRUE)
cas2 = casesinlocality(obj = casa, locality = "AP1")
tw = getTweet(city = c(330455))
clima = getWU(stations = 'SBRJ', var="tmin")
head(mergedata(cases = cas,tweet = tw, climate = clima))
head(mergedata(cases = cas2,tweet = tw, climate = clima))
head(mergedata(tweet = tw, climate = clima))
head(mergedata(cases = cas, climate = clima))
head(mergedata(cases = cas2, climate = clima))
head(mergedata(tweet = tw, cases = cas))
head(mergedata(tweet = tw, cases = cas2))
}

