% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tentative-functions.R
\name{mergedata2}
\alias{mergedata2}
\title{Define conditions to issue a 4 level alert Green-Yellow-Orange-Red for a city.}
\usage{
mergedata2(cases = c(), tweet = c(), climate = c(), ini = 200952)
}
\arguments{
\item{city}{city's geocode (6 or 7 digits).}

\item{pars}{}

\item{datasource}{PosgreSQL connection}

\item{writedb}{TRUE if it should write into the database, default is FALSE.}

\item{sefinal}{if given, it stops at that week}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario. For most places weather stations are provided in the database and choice
is authomatic according to data quality. But it can also be provided manually (not implemented).
}
\examples{
# Parameters for the model
criteria = list(crity = c("temp_min > tcrit", 3, 1),
crito = c("p1 > 0.95 & inc > preseas & temp_min >= tcrit", 3, 1),
critr = c("inc > inccrit", 2, 2))
gtdist="normal"; meangt=3; sdgt = 1.2
pars.RJ <- NULL
pars.RJ <- list(pdig = c(2.997765,0.7859499),tcrit=22, inccrit = 100, preseas=8.28374162389761, posseas = 7.67878514885295, legpos="bottomright")
Running the model:
res <- update.alerta(city = 3302205, pars = pars.RJ, crit = criteria, 
datasource = con)
tail(res$data)
}
