% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{nafill}
\alias{nafill}
\title{methods to substitute NAs. Use the function na.approx from package zoo.}
\usage{
nafill(v, rule, maxgap = 4, verbose = F)
}
\arguments{
\item{v}{vector with missing elements.}

\item{rule}{rule for filling the missing cells. "zero" just fills them with 0; "linear"
interpolate using zoo::na.approx. In this case, the tails are not filled. If "arima", then it interpolates using
linear and extrapolates using arima (calling AlertTools::temp.predict)}

\item{maxgap}{maximum number of consecutive NAs to fill. Longer gaps will be left unchanged. Only works for rule = "zero"
or "linear"}

\item{verbose}{Whether to emit information about missing values.}
}
\value{
vector
}
\description{
collection of imputation procedures
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Interpolation:
v <- c(1,2,3,NA,5,6,NA,NA,9,10,NA,NA)
nafill(v, rule = "zero")
nafill(v, rule = "linear")
# Inter using linear and Extrapolation using arima
cliSBCB <- getWU(station = "SBCB")
summary(cliSBCB)
cliSBCB <- getWU(station = "SBCB") \%>\%
          mutate(nafill("temp_min", rule = "arima"))
\dontshow{\}) # examplesIf}
}
