% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_infodengue.R
\name{pipe_infodengue_22}
\alias{pipe_infodengue_22}
\title{pipeline used by infodengue}
\usage{
pipe_infodengue_22(
  state,
  cid10 = "A90",
  datarelatorio,
  finalday = Sys.Date(),
  iniSE = 201001,
  nowcasting = "none",
  narule = NULL,
  writedb = FALSE,
  datasource = con,
  recept_rules,
  orange_rule = 3,
  completetail = NA,
  dataini = "sinpri"
)
}
\arguments{
\item{cid10}{default is A90 (dengue). Chik = A92.0, Zika = A92.8}

\item{datarelatorio}{epidemiological week}

\item{finalday}{if provided, uses only disease data reported up to that day. 
class must be Date.}

\item{iniSE}{first date of the disease data. Default = 201501. Minimum = 201001.}

\item{nowcasting}{"bayesian" for the dynamic model; "none" for not doing nowcast (default)}

\item{narule}{how to treat missing climate data. Do nothing (default), "zero" fills 
with zeros, "linear" for linear interpolation, "arima" for inter and extrapolation.}

\item{writedb}{TRUE if it should write into the database, default is FALSE.}

\item{datasource}{posgreSQL connection to project's database}

\item{orange_rule}{number of weeks with Rt > 1 above which orange is raised}

\item{completetail}{if sinan data is older than final_day, fill in the tail 
with NA (default) or 0.}

\item{dataini}{"notif" (default) or "sinpri"}

\item{cities}{In general, a vector of 7-digit geocodes. If it is a data.frame containing geocodes
and all parameters, these will replace the database's parameters.}

\item{yellow_rule}{decision rule for yellow alert (climate receptivity)}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
wrap of functions used by Infodengue.
}
\examples{
d <- pipe_infodengue_22(state = "Acre", cid10 = "A90", iniSE = 201801,
finalday= as.Date("2021-08-12"),nowcasting="none", completetail = 0, 
dataini = "sinpri", recept_rules = AlertTools::regras)
restab <- tabela_historico(d)
}
