% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{pipe_infodengue_intra}
\alias{pipe_infodengue_intra}
\title{4 level alert Green-Yellow-Orange-Red for partitions within municipalities.}
\usage{
pipe_infodengue_intra(
  city,
  locs,
  datarelatorio,
  cid10 = "A90",
  iniSE = 201001,
  delaymethod = "none",
  narule = "arima",
  finalday = Sys.Date(),
  dataini = "sinpri",
  datasource = con
)
}
\arguments{
\item{city}{geocode (one city at a time). Currently only works for Rio.}

\item{locs}{subset of localities. Currently not used.}

\item{datarelatorio}{last epidemiological week (format = 201401)}

\item{cid10}{default is A90 (dengue). Chik = A920.}

\item{iniSE}{first epidemiological week}

\item{delaymethod}{"none" (Default) or "bayesian"}

\item{narule}{how to fill climate missing data (arima is the only option)}

\item{dataini}{"sinpri" or "notif". Default = "sinpri"}

\item{datasource}{name of the sql connection.}
}
\value{
list with an alert object for each APS.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
alerio <- pipe_infodengue_intra(city = 3304557, datarelatorio=202113, 
delaymethod="none", cid10 = "A90", dataini = "sinpri")
ale.chik <- pipe_infodengue_intra(city = 3304557, datarelatorio = 202108, 
iniSE = 201001, cid10 = "A920", dataini = "sinpri", delaymethod = "bayesian")
}
