% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{plot.alerta}
\alias{plot.alerta}
\title{Plot the time series of warnings.}
\usage{
\method{plot}{alerta}(obj, var, cores = c("#0D6B0D", "#C8D20F", "orange",
  "red"))
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{var}{to be ploted in the graph, usually cases when available.}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}
}
\value{
a plot
}
\description{
Function to plot the output of
}
\examples{
tw <- getTweet(city = c(330455), lastday = "2014-03-10")
clima <- getWU(stations = 'SBRJ', var="tmin", finalday = "2014-03-10")
d<- mergedata(tweet = tw, climate = clima)
critgy <- c("tmin > 22 | tweet > 10", 3)
crityg <- c("tmin <= 22 & tweet <= 10", 3)
alerta <- twoalert(d, gy = critgy, yg = crityg)
plot.alerta(alerta, "tmin")
}

