% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{plot.inla.re}
\alias{plot.inla.re}
\title{plot delay and time random effects}
\usage{
plot.inla.re(outputRE, xlab)
}
\arguments{
\item{outputRE}{random effect components of the object created by the fitDelay.inla function}
}
\value{
graphs
}
\description{
Plot the random effects of the delay model fitted using fitDelay.inla()
}
\examples{
res = delaycalc(dados)
outp<-fitDelay.inla(res)
par(mfrow=c(2,1),mar=c(4,4,2,2))
plot.inla.re(outp$out$summary.random$Time, xlab="semana epidemiologica")
plot.inla.re(outp$out$summary.random$Delay, xlab="semana de atraso")
}
