% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{plot_alertaRio}
\alias{plot_alertaRio}
\title{Plot the time series of alerts for each APS. Specific for the Rio de Janeiro alert model.}
\usage{
plot_alertaRio(ale, aps, var = "casos", cores = c("#0D6B0D", "#C8D20F",
  "orange", "red"), ini = 201001, fim = 202001, ylab = var, yrange,
  salvar = FALSE, nome.fig = "grafico", datasource = con)
}
\arguments{
\item{ale}{object created by alertaRio()}

\item{aps}{aps to be plotted. If missing, figures will be created for all aps (health districts)}

\item{var}{variable to be plotted, usually "cases", or "inc".}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}

\item{ini}{first epidemiological week. If not stated, use the first one available}

\item{fim}{last epidemiological week. If not stated, use the last one available}

\item{ylab}{y axis label}

\item{yrange}{y axis range}

\item{salvar}{TRUE to save the figure (default is FALSE)}

\item{nome.fig}{figure name (default is "grafico")}
}
\value{
a plot
}
\description{
Function to plot the output of alertaRio().
}
\examples{
# Parameters for the model
alerio2 <- alertaRio(naps = 0:9, se=201804, delaymethod="fixedprob", cid10 = "A90")
plot_alertaRio(alerio2, var = "inc")
}
