% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{prob.inc}
\alias{prob.inc}
\title{posterior distribution of the incidence}
\usage{
prob.inc(obj, plotar = T)
}
\arguments{
\item{obj}{created by the fitDelay.inla function}
}
\value{
table with mean, median, 2.5% and 97.5% incidence.
}
\description{
predicted incidence using fitDelay.inla().
}
\examples{
res = delaycalc(dados)
outp<-fitDelay.inla(res)
delay <- prob.inc(outp)
}
\author{
Leo Bastos
}
