% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{read.parameters}
\alias{read.parameters}
\title{Legacy alert-parameter reader}
\usage{
read.parameters(cities, cid10 = "A90", datasource)
}
\arguments{
\item{cities, cid10, datasource}{Legacy arguments.}
}
\description{
Legacy alert-parameter reader
}
