% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Rt}
\alias{Rt}
\title{Legacy Rt estimator}
\usage{
Rt(
  obj,
  count = "casos",
  gtdist,
  meangt,
  sdgt,
  CI = "beta",
  alpha = 0.95,
  a0 = 2,
  b0 = 3
)
}
\arguments{
\item{obj, count, gtdist, meangt, sdgt, CI, alpha, a0, b0}{Legacy arguments.}
}
\description{
Legacy Rt estimator
}
