% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{SE2date}
\alias{SE2date}
\title{Return the first day of the Epidemiological Week}
\usage{
SE2date(se)
}
\arguments{
\item{se}{numeric vector with epidemiological weeks, for example 201420.}
}
\value{
data.frame with SE and first day.
}
\description{
Return the first day of the Epidemiological Week
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
SE2date(se=202612)
SE2date(se = c(202001:202209))
\dontshow{\}) # examplesIf}
}
