% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{setCriteria}
\alias{setCriteria}
\title{Legacy alert-rule constructor}
\usage{
setCriteria(rule = NULL, values = NULL, delays = 3)
}
\arguments{
\item{rule, values, delays}{Legacy arguments.}
}
\description{
Legacy alert-rule constructor
}
