% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{setOrange}
\alias{setOrange}
\title{Raise orange alert if sustained transmission is detected.}
\usage{
setOrange(obj, pvalue = 0.9, lag = 3)
}
\arguments{
\item{obj}{object created by the Rt function.}

\item{pvalue}{probability of wrongly rejecting the hypothesis Rt > 1}

\item{lag}{count the number of weeks within the last lag weeks with conditions = TRUE}
}
\value{
data.frame with the week condition and the number of weeks within the
last lag weeks with conditions = TRUE
}
\description{
"Orange" is defined by effective reproductive number greater than 1.
Rt must be computed before using this function.
}
\examples{
res <- getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
resd <- aggrbylocality(d = res, locality="AP1") # Rio de Janeiro
resfit <- adjustIncidence(se = res$SE, y = res$casos)
rtnorm<-Rt(obj = resfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
ora = setOrange(obj = rtnorm, pvalue = 0.9, lag= 3)
head(ora)
x = 1:length(ora$SE)
plot(x, ora$Rt, type="l", xlab= "weeks", ylab = "Rt")
lines(x, ora$upr, lty = 3)
lines(x, ora$lwr, lty = 3)
abline(h = 1, col =2)
points(x[ora$oweek==1], ora$Rt[ora$oweek==1], col="orange", pch=16)
}

