% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{setRed}
\alias{setRed}
\title{Raise red alert if incidence reaches above a threshold.}
\usage{
setRed(obj, pop, ccrit = 100, lag = 3)
}
\arguments{
\item{obj}{case count dataset generated by the getcase (ajusted or not).}

\item{pop}{population of the area.}

\item{ccrit}{incidence threshold (per 100.000).}

\item{lag}{count the number of weeks within the last lag weeks with conditions = TRUE.}

\item{adjust}{TRUE if adjusting incidence is required. FALSE if not.}
}
\value{
data.frame with the week condition and the number of weeks within the
last lag weeks with conditions = TRUE
}
\description{
"Red" indicates high dengue incidence, defined by a threshold provided
by the user. The WHO recommends 300 per 100.000.
}
\examples{
res = getCases(city = c(330455), withdivision = TRUE) # Rio de Janeiro
resd = aggrbylocality(d = res, locality="AP1") # Rio de Janeiro
resfit<-adjustIncidence(se = resd$SE, y = resd$casos)
red = setRed(obj = resfit, pop = 30000, ccrit = 10, lag=3)
tail(red)
x = 1:length(red$SE)
plot(x, red$inc, type="l", xlab= "weeks", ylab = "incidence")
abline(h = 10, col =2)
points(x[red$rweek==1], red$inc[red$rweek==1], col="red", pch=16)
}

