% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{setWUstation}
\alias{setWUstation}
\title{set meteorological stations}
\usage{
setWUstation(st, UF, datasource)
}
\arguments{
\item{st}{data.frame containing municipio_geocodigo, primary_station,
secondary_station}

\item{UF}{name of the state.Ex. "Rio de Janeiro"}

\item{datasource}{connection to the project database}
}
\description{
Set primary and secondary meteorological stations associated
with one or more cities of the same state
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# NOT RUN
wudata = data.frame(municipio_geocodigo = 3107802, primary_station = "SBIP",
secondary_station = "SBGV")
setWUstation(wudata, UF = "Minas Gerais")
getWUstation(cities =wudata$municipio_geocodigo)
\dontshow{\}) # examplesIf}
}
