% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{setYellow}
\alias{setYellow}
\title{Define conditions to issue an alert of level Yellow.}
\usage{
setYellow(temp, se, tempcrit, lag = 3)
}
\arguments{
\item{temp}{time series of temperature.}

\item{se}{sequence of epidemiological weeks.}

\item{tempcrit}{critical temperature.}

\item{lag}{count the number of weeks within the last lag weeks with conditions = TRUE.}
}
\value{
data.frame with the week condition and the number of weeks within the
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected.
}
\examples{
clima = getWU(stations = "SBRJ", var = "tmin")
yellow = setYellow(temp = clima$tmin, se = clima$SE,  tempcrit = 22)
head(yellow)
x = 1:length(yellow$SE)
plot(x, yellow$temp, type="l", xlab= "weeks", ylab = "Temperature")
abline(h = 22, col =2)
points(x[yellow$yweek==1], yellow$temp[yellow$yweek==1], col="yellow", pch=16)
}

