% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{sevendigitgeocode}
\alias{sevendigitgeocode}
\title{convert 6 to 7 digits geocodes.}
\usage{
sevendigitgeocode(dig)
}
\arguments{
\item{dig}{Six- or seven-digit municipality geocode.}
}
\value{
7 digits municipality geocode.
}
\description{
calculates the verification digit of brazilian municipalities. Required
to convert 6 digits to 7 digits geocodes.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
sevendigitgeocode(330455)
sevendigitgeocode(3304557)
sevendigitgeocode(261153)
\dontshow{\}) # examplesIf}
}
