% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{tabela.historico}
\alias{tabela.historico}
\title{Write the alert object into the database.}
\usage{
tabela.historico(obj, versao = Sys.Date())
}
\arguments{
\item{obj}{object created by the pipeline.}

\item{versao}{Default is current's date}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe.infodengue(cities = cidades$municipio_geocodigo, cid10 = "A90", 
finalday= "2013-01-10")
restab <- tabela.historico(res)
}
