% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tabela_historico}
\alias{tabela_historico}
\title{Legacy history converter}
\usage{
tabela_historico(
  obj,
  iniSE,
  lastSE,
  type = "all",
  versao,
  parameters = if (inherits(obj, "alerttools_result")) obj$parameters else attr(obj,
    "parameters")
)
}
\arguments{
\item{obj, iniSE, lastSE, type, versao, parameters}{Legacy arguments.}
}
\description{
Legacy history converter
}
