% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{update.alerta}
\alias{update.alerta}
\title{Define conditions to issue a 4 level alert Green-Yellow-Orange-Red for any city or region.}
\usage{
\method{update}{alerta}(city, region, state, pars, crit, writedb = FALSE,
  datasource, sefinal)
}
\arguments{
\item{city}{city's geocode (6 or 7 digits).}

\item{region}{full name of 'regional' or state (same name present in the database).}

\item{pars}{list of parameters for the alerta, defined in config.R}

\item{crit}{criteria for the alert colors, defined in config.R}

\item{writedb}{TRUE if it should write into the database, default is FALSE.}

\item{sefinal}{if given, it stops at that week}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
res <- update.alerta(city = 330240, pars = pars.RJ, crit = criteria, datasource = con)
res <- update.alerta(region = names(pars.RJ), pars = pars.RJ, crit = criteria, datasource = con,sefinal=201613)
tail(res$data)
}

