% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{updateDelayModel}
\alias{updateDelayModel}
\title{Update notification delay model.}
\usage{
updateDelayModel(
  cities,
  ufs,
  period,
  datasource,
  plotar = FALSE,
  write,
  verbose = FALSE,
  regional = FALSE
)
}
\arguments{
\item{cities}{geocode of one of more cities.}

\item{ufs}{list of ufs (full name as in the database). Required even if only a subset of cities is the target}

\item{period}{range of dates for the analysis. Format: c("2010-01-01","2015-12-31"). 
Default is the whole period available.}

\item{datasource}{sql connection.}

\item{plotar}{if TRUE, show plot of the fitted model.}

\item{write}{TRUE if result.}

\item{regional}{TRUE, if model should be fitted at the regional level too}
}
\value{
object with a summary of the analysis and suggestion of best model.
}
\description{
Apply the fitDelayModel to all or a subset of cities.
}
\examples{
con <- DenguedbConnect()
par(mfrow=c(2,1))
res<-updateDelayModel(cities=c(330240, 330045), period=c("2013-01-01","2016-01-01"), plotar=TRUE, ufs = "Rio de Janeiro", datasource=con)
res<-updateDelayModel(ufs="Rio de Janeiro", period=c("2013-01-01","2016-01-01"), datasource=con)
res<-updateDelayModel(ufs="Rio de Janeiro", period=c("2013-01-01","2016-01-01"), regional=TRUE, datasource=con)
}
