% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{weeks_above_threshold}
\alias{weeks_above_threshold}
\title{Counts the number of successive weeks with incidence above a given threshold}
\usage{
weeks_above_threshold(d, var = "inc", varname = "wat", threshold, maxw = 52)
}
\arguments{
\item{d}{dataset with data to feed the alert. It is recommended the adjustment
of incidence first}

\item{var}{variable used for computation (default = inc)}

\item{varname}{name of the output variable. Default : wat}

\item{threshold}{incidence threshold. Either a number to be applied to all , or 
a data.frame with two columns: geocode, threshold.}

\item{maxw}{max number of weeks (default = 52)}
}
\value{
returns the data, and the number of weeks with transmission.
}
\description{
Indicator used for raising the red alert
}
\examples{
# Parameters of the alert model (requires connection)
mun <- getCidades(uf = "Acre")
casos <- getCases(mun$municipio_geocodigo, cid10 = "A90", 
dataini = "sinpri", lastday = "2018-05-01") \%>\% mutate(inc = casos/pop*100000)
pars <- read.parameters(mun$municipio_geocodigo)
thre <- pars[, c("municipio_geocodigo","limiar_epidemico")]
casos <- casos \%>\% weeks_above_threshold(var = "inc", threshold = thre, varname = "z")
boxplot(casos$z ~ casos$cidade, type = "s", main = "weeks_above_threshold")
# for the pipeline, we use maxw = 2
casos <- casos \%>\% weeks_above_threshold(var = "casos", maxw = 2)
}
