% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alerta}
\alias{write.alerta}
\title{Write the alert into the database.}
\usage{
write.alerta(obj, ini, end, write = "no", replacelast = 26,
  version = Sys.Date(), NAaction = "omitcasesmedNA")
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{ini}{first epidemiological week of the period to be written.}

\item{end}{last epidemiological week of the period to be written.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}

\item{replacelast}{weeks to replace the ones present in the database. replacelast = 26 
updates the db for the last 6 months. replacelast = 9999 replace all db.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
tw = getTweet(city = 330070, datasource = con) 
cli = getWU(stations = 'SBCB', datasource=con)
cas = getCases(city = 330070, withdivision = FALSE, datasource=con)
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = cli)
crity <- c("temp_min > 22", 3, 3)
crito <- c("p1 > 0.9", 3, 3)
critr <- c("inc > 100", 3, 3)
alerta <- fouralert(d, cy = crity, co = crito, cr = critr, pop=1000000)
res <- write.alerta(alerta, write="db", replacelast = 5)
tail(res)
}

