% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alertaRio}
\alias{write.alertaRio}
\title{Write the Rio de janeiro alert into the database.}
\usage{
write.alertaRio(obj, write = "no", version = Sys.Date())
}
\arguments{
\item{obj}{object created by the alertRio function and contains alerts for each APS.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
alerio2 <- alertaRio(naps = c(1,2), datasource=con)
res <- write.alertaRio(alerio2, write="db")
tail(res)
}

