% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{write.parameters}
\alias{write.parameters}
\title{City's parameterization.}
\usage{
write.parameters(params, overwrite = FALSE, senha)
}
\arguments{
\item{params}{vector of the names of the params to be inserted in the table. Default: 
params = data.frame("municipio_geocodigo","limiar_preseason", "limiar_posseason","limiar_epidemico"
"varcli","clicrit","cid10","codmodelo"). It can be a subset of the default.}
}
\value{
the new line in the parameters table
}
\description{
Write the alert parameters for each city into the database, to be used in the update.alert. 
Currently, the parameters are:"codigo_estacao_wu", "limiar_preseason", "limiar_posseason",
"limiar_epidemico, "estacao_wu_sec".  City must be already in the regionais table.
}
\examples{
newpars = data.frame(municipio_geocodigo = 0000000, cid10 = "A90", limiar_preseason=2, limiar_posseason=2,
varcli = "temp_min", clicrit = 20)
res = write.parameters(newpars)
}
