% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{write_alerta}
\alias{write_alerta}
\title{Legacy alert-result writer}
\usage{
write_alerta(
  d,
  writetofile = FALSE,
  datasource,
  arq = "output.sql",
  conflict = c("update", "error", "ignore")
)
}
\arguments{
\item{d, writetofile, datasource, arq, conflict}{Legacy arguments.}
}
\description{
Legacy alert-result writer
}
