% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write_alerta_local}
\alias{write_alerta_local}
\title{Write historico_alerta into the local SQLite database.}
\usage{
write_alerta_local(d, datasource = con)
}
\arguments{
\item{d}{object created by tabela_historico()}

\item{datasource}{SQLite conn}
}
\value{
the same data.frame from the input
}
\description{
Function to write the output of the pipeline into the local SQLite database. 
The input must be the object created by the function historico.alerta.
}
\examples{
NOT USE: con <- dbConnect(RSQLite::SQLite(), "../../AlertaDengueAnalise/mydengue.sqlite")
# Parameters for the model 
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo, cid10 = "A90", 
finalday= "2021-07-12",nowcasting="none", dataini = "sinpri")
restab <- tabela_historico(res)
# NOT RUN 
t1 <- Sys.time()
write_alerta_local(restab)
t2 <- Sys.time() - t1
}
