% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write_alertaRio}
\alias{write_alertaRio}
\title{Write the Rio de janeiro alert into the database.}
\usage{
write_alertaRio(obj, write = "no", datasource = con, version = Sys.Date())
}
\arguments{
\item{obj}{object created by the tabela_historico_intra function and contains alerts for each locality.}

\item{version}{default is the current date}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
res <- write_alertaRio(alerio2, write="no")
tail(res)  
}
