% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{write_parameters}
\alias{write_parameters}
\title{Legacy alert-parameter writer}
\usage{
write_parameters(
  city,
  cid10,
  params,
  overwrite = FALSE,
  datasource,
  conflict = c("ignore", "update", "error")
)
}
\arguments{
\item{city, cid10, params, overwrite, datasource, conflict}{Legacy arguments.}
}
\description{
Legacy alert-parameter writer
}
