% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{adjustIncidence}
\alias{adjustIncidence}
\title{Adjust incidence data correcting for the notification delay.}
\usage{
adjustIncidence(y, dist = "lognormal", meanlog = 2.5016, sdlog = 1.1013)
}
\arguments{
\item{y}{time series of cases to be adjusted.}

\item{fun}{function describing the proportion reported as a function of time passed.
Currently only the accumulated lognormal function is implemented .}

\item{pars}{list of parameters for the function fun.}
}
\value{
data.frame with p (proportion reported), lambda ( expected numbero f cases not yet reported),
inc (expected true number of cases to be reported) incICmin and incICmax (95percent confidence interval for the true incidence)
}
\description{
Often, there is a delay between symptom onset and notification. This function
adjust the time series of reported cases adding the cases that will be reported in the future.
A function describing the proportion known as a function of time passed is required.
}
\examples{
res = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
res$fit<-adjustIncidence(res$casos)
tail(res)
plot(st$casos[232:262],type="l")
lines(st$casosm[232:262],col=2)
}

