% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{adjustIncidence}
\alias{adjustIncidence}
\title{Adjust incidence data correcting for the notification delay.}
\usage{
adjustIncidence(se, y, dist = "lognormal", meanlog = 2.5016,
  sdlog = 1.1013, prop)
}
\arguments{
\item{y}{time series of cases to be adjusted.}

\item{dist}{function describing the proportion reported as a function of time passed.
Currently only the accumulated lognormal function is implemented .}

\item{meanlog}{mean of the lognormal function.}

\item{prop}{alternatively, a vector with the proportion typed per week (not implemented yet).}

\item{meanlog}{sdev of the lognormal function.}
}
\value{
data.frame with p (proportion reported), lambda ( expected numbero f cases not yet reported),
inc (expected true number of cases to be reported) incICmin and incICmax (95percent confidence interval for the true incidence)
}
\description{
Often, there is a delay between symptom onset and notification. This function
adjust the time series of reported cases adding the cases that will be reported in the future.
A function describing the proportion known as a function of time passed is required.
}
\examples{
res = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
head(res)
resfit<-adjustIncidence(se = res$SE, y = res$casos)
tail(resfit)
plot(tail(resfit$casos,n=30),type="l",ylab="cases",xlab="weeks")
lines(tail(resfit$tcasesmed,n=30),col=2)
lines(tail(resfit$tcasesmin,n=30),col=2,lty=3)
lines(tail(resfit$tcasesmax,n=30),col=2,lty=3)
legend(12,20,c("notified cases","+ to be notified cases"),lty=1, col=c(1,2),cex=0.7)
}

