% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{adjustIncidence}
\alias{adjustIncidence}
\title{Adjust incidence data correcting for the notification delay.}
\usage{
adjustIncidence(obj, pdig = plnorm((1:20) * 7, 2.5016, 1.1013))
}
\arguments{
\item{obj}{data.frame with crude weekly cases (not adjusted). This data.frame comes from the getCases
function (if withdivision = FALSE), of getCases followed by casesinlocality (if dataframe is available
per bairro)}

\item{pdig}{vector of probability of been typed in the database up to 1, 2, 3, n, weeks after symptoms onset.
The length of the vector corresponds to the maximum delay. After day, it is assumed that p = 1. The default
was obtained from Rio de Janeiro.}
}
\value{
data.frame with pdig (proportion reported), median and 95percent confidence interval for the 
predicted cases-to-be-notified)
}
\description{
Often, there is a delay between symptom onset and notification. This function 
adjust the time series of reported cases by adding the cases that will be reported in the future.
It requires knowing the probability of notification per week passed. This function assumes a stationary
notification process, there is, no influence of covariates or any temporal inhomogeneity.
}
\examples{
res = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
head(res)
resfit<-adjustIncidence(obj=res)
tail(resfit)
plot(tail(resfit$casos,n=30),type="l",ylab="cases",xlab="weeks")
lines(tail(resfit$tcasesmed,n=30),col=2)
lines(tail(resfit$tcasesmin,n=30),col=2,lty=3)
lines(tail(resfit$tcasesmax,n=30),col=2,lty=3)
legend(12,20,c("notified cases","+ to be notified cases"),lty=1, col=c(1,2),cex=0.7)
}

