% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{alertaRio}
\alias{alertaRio}
\title{4 level alert Green-Yellow-Orange-Red for Rio de Janeiro.}
\usage{
alertaRio(naps = 0:9, pars = list(meanlog = 2.5016, sdlog = 1.1013, crity =
  c("temp_min > 22", 3, 3), crito = c("p1 > 0.9", 3, 3), critr = c("inc > 100",
  3, 2)), datasource)
}
\arguments{
\item{naps}{subset of vector 0:9 corresponding to the id of the APS. Default is all of them.}

\item{pars}{parameters of the alert.}

\item{datasource}{it is the name of the sql connecation.}
}
\value{
list with an alert object for each APS.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
alerio2 <- alertaRio(naps = c(0,1), datasource=con)
names(alerio2)
}

