% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{alertaRio}
\alias{alertaRio}
\title{4 level alert Green-Yellow-Orange-Red for Rio de Janeiro.}
\usage{
alertaRio(naps = 0:9, pars, crit, datasource, se, cid10 = "A90",
  verbose = TRUE, delaymethod = "fixedprob", dataini = "notific")
}
\arguments{
\item{naps}{subset of vector 0:9 corresponding to the id of the APS. Default is all of them.}

\item{pars}{parameters of the alert.}

\item{datasource}{it is the name of the sql connection.}

\item{se}{last epidemiological week (format = 201401)}

\item{cid10}{default is A90 (dengue). Chik = A920.}

\item{verbose}{FALSE}

\item{delaymethod}{atribbute of adjuntincidence. "fixedprob" or "bayesian"}

\item{dataini}{"notific" if use notification date to calculate incidence or "sinpri" if uses date of first symptoms}
}
\value{
list with an alert object for each APS.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
params <- list(pdigChik = c(2.687797, 1.362566), tcrit=22, inccrit=100, preseas = 14.15, posseas = 18)
criter = list(
     crity = c("temp_min > tcrit & casos > 0", 3, 1),
     crito = c("p1 > 0.95 & temp_min >= tcrit", 3, 1),
     critr = c("inc > inccrit & casos > 5", 2, 2))
alerionot <- alertaRio(pars=params, cid = "A920", crit = criter, datasource=con, se=201918,
delaymethod="fixedprob",verbose=FALSE)
aleriosinpri <- alertaRio(pars=params, cid = "A920", crit = criter, datasource=con, se=201918,
delaymethod="fixedprob",verbose=FALSE,dataini = "sinpri")
}
