% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{bayesnowcasting}
\alias{bayesnowcasting}
\title{Correct incidence data with notification delay (nowcasting).}
\usage{
bayesnowcasting(d, Dmax = 10, nowSE, interacao = TRUE)
}
\arguments{
\item{Dmax}{for the "bayesian" method. Maximum number of weeks that is modeled}

\item{nowSE}{week of the nowcasting (ex. 202110).}

\item{interacao}{TRUE (default) se o modelo tiver o termo de interacao efeito-atraso}

\item{obj}{data.frame with individual cases, containing columns municipio_geocodigo, dt_notific, dt_sin_pri, dt_digita}
}
\value{
data.frame with median and 95percent confidence interval for the 
predicted cases-to-be-notified)
}
\description{
This function estimate the Bayesian nowcast (new version)
}
\examples{
# bayesian
dd <- getdelaydata(cities=2304400, nyears=1, cid10="A90", datasource=con)
resfit<-bayesnowcasting(dd, nowSE = 202111)
resfit
}
