% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{bayesnowcasting}
\alias{bayesnowcasting}
\title{Correct incidence data with notification delay (nowcasting).}
\usage{
bayesnowcasting(d, Dmax = 10, nowSE, nweeks = 50, interacao = TRUE, tweet = F)
}
\arguments{
\item{Dmax}{for the "bayesian" method. Maximum number of weeks that is nowcasted}

\item{nowSE}{week of the nowcasting (ex. 202110).}

\item{nweeks}{number of weeks used for model calibration, default is 50.}

\item{interacao}{TRUE (default) to include in the model the delay-time interaction term}

\item{tweet}{FALSE (default). TRUE to include tweet in the model}

\item{obj}{data.frame with individual cases, containing columns municipio_geocodigo, dt_notific, dt_sin_pri, dt_digita}
}
\value{
data.frame with median and 95percent confidence interval for the 
predicted cases-to-be-notified)
}
\description{
This function estimate the Bayesian nowcast (new version)
}
\examples{
dados <- getdelaydata(cities=3304557, nyears=1, cid10="A90", 
lastday = as.Date("2023-12-30"), datasource=con)  # Not run without connection
resfitcIsT<-bayesnowcasting(dados, nowSE = 202352)
resfitcIcT<-bayesnowcasting(dados, nowSE = 201945, tweet = TRUE)
}
