% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{data2SE}
\alias{data2SE}
\title{Define Epidemiological Week}
\usage{
data2SE(days, format = "\%d/\%m/\%Y")
}
\arguments{
\item{format}{date format}

\item{date}{string vector with dates to be converted}
}
\value{
data.frame with the epidemiological weeks. Only works for years >= 2010.
}
\description{
Find to which epidemiological week belongs a given day (only >=2010)
}
\examples{
data2SE("01-02-2010",format="\%d-\%m-\%Y")
data2SE(c("03-04-2013","07-01-2014"),format="\%d-\%m-\%Y")
}

