% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{daySEday}
\alias{daySEday}
\title{Return the first day of the Epidemiological Week and vice-versa}
\usage{
daySEday(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{numeric vector with epidemiological weeks , format 201945, or date}
}
\value{
data.frame with SE and first day.
}
\description{
Return the first day of the Epidemiological Week and vice-versa
}
\examples{
daySEday(x=201812)
daySEday(x = c(202041:202104))
daySEday(x = c("2015-12-23", "2015-10-23", "2022-10-16"))
}
