% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{episem}
\alias{episem}
\title{Define Epidemiological Week}
\usage{
episem(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{date to be converted (class Date)}

\item{separa}{symbol between year and week}
}
\value{
epidemiological week
}
\description{
Find to which epidemiological week belongs a given day (Oswaldo's function)
}
\examples{
episem(x= as.Date("2015-01-01", format="\%Y-\%m-\%d"))
}

