% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{fitDelayModel}
\alias{fitDelayModel}
\title{Fit lognormal model to the notification delay}
\usage{
fitDelayModel(cities, period, plotar = TRUE, cid10 = "A90", datasource,
  verbose = TRUE, inidate = "dt_notific")
}
\arguments{
\item{cities}{geocode of one of more cities. If more than one city, a single model is fitted to the whole dataset.}

\item{period}{range of dates for the analysis. Format: c("2010-01-01","2015-12-31"). 
Default is the whole period available.}

\item{plotar}{if TRUE, show plot of the fitted model}

\item{cid10}{CID 10 code of the disease. Dengue = "A90" is default, Chik = "A920"}

\item{datasource}{sql connection}

\item{cities}{geocode of one of more cities. If more than one city, a single model is fitted to the whole dataset.}

\item{period}{range of dates for the analysis. Format: c("2010-01-01","2015-12-31"). 
Default is the whole period available.}

\item{dateini}{either dt_notific(default) or dt_sin_pri}

\item{datasource}{sql connection}

\item{plotar}{if TRUE, show plot of the fitted model}
}
\value{
object with a summary of the analysis and suggestion of best model

object with a summary of the analysis and suggestion of best model
}
\description{
Fit lognormal model to the notification delay and return the parameters. See
details in doi: http://dx.doi.org/10.1101/046193

Fit lognormal model to the notification delay and return the parameters.
}
\examples{
con <- DenguedbConnect()
res1<-fitDelayModel(cities=330240, datasource=con)
res<-fitDelayModel(cities=330240, period=c("2013-01-01","2016-01-01"), datasource=con)
# Parameters are
list(meanlog=res$icoef[1], sdlog=exp(res$icoef[2]))
con <- DenguedbConnect()
res1<-fitDelayModel(cities=330240, datasource=con)
res<-fitDelayModel(cities=330240, period=c("2013-01-01","2016-01-01"), datasource=con)
res1<-fitDelayModel(cities=3304557, cid10="A920",datasource=con)
# Parameters are
list(meanlog=res$icoef[1], sdlog=exp(res$icoef[2]))
}
