% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{fouralert}
\alias{fouralert}
\title{Define conditions to issue a 4 level alert Green-Yellow-Orange-Red.}
\usage{
fouralert(obj, pars, crit, pop, miss = "last")
}
\arguments{
\item{obj}{dataset from the mergedata function.}

\item{pars}{list with criteria for each color. See example.}

\item{missing}{how missing data is treated. "last" if last value is repeated. 
It is currently the only option.}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
tw = getTweet(city = 4102000, datasource = con) 
cli = getWU(stations = 'SBCA', datasource = con)
cas = getCases(city = 4102000, datasource=con)
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = cli)
crit <- list(pdig = c(2.5016,1.1013), tcrit=22, inccrit=100, preseas = 14.15, posseas = 18,
              crity = c("temp_min > tcrit | (temp_min < tcrit & inc > preseas)", 3, 0),
              crito = c("p1 > 0.9 & inc > preseas", 2, 2),
              critr = c("inc > inccrit", 1, 2))
ale <- fouralert(d, pars = crit, pop = 1000000)
# For a more useful output
res <- writeAlerta(ale)
tail(res)
}

