% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{fouralert}
\alias{fouralert}
\title{Define conditions to issue a 4 level alert Green-Yellow-Orange-Red.}
\usage{
fouralert(obj, cy, co, cr, pop, miss = "last")
}
\arguments{
\item{obj}{dataset from the mergedata function.}

\item{cy}{conditions for yellow.}

\item{co}{conditions for orange.}

\item{cr}{conditions for red.}

\item{missing}{how missing data is treated. "last" if last value is repeated. 
It is currently the only option.}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
tw = getTweet(city = c(330455), datasource = "data/tw.rda") 
cli = getWU(stations = 'SBRJ', datasource="data/WUdata.rda")
cas = getCases(city = c(330455), withdivision = FALSE, datasource="data/sinan.rda")
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = clima)
crity <- c("temp_min > 22", 3, 3)
crito <- c("p1 > 0.9", 3, 3)
critr <- c("inc > 100", 3, 3)
alerta <- fouralert(d, cy = crity, co = crito, cr = critr, pop=1000000)
plot.alerta(alerta, var="casos", ylab="casos")
# For a more useful output
res <- write.alerta(alerta)
tail(res)
}

