% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{fouralert}
\alias{fouralert}
\title{Define conditions to issue a four level alert Green-Yellow-Orange-Red.}
\usage{
fouralert(obj, crit, miss = "last")
}
\arguments{
\item{obj}{dataset with data to feed the alert, containing the variables specified in crit.}

\item{crit}{criteria for the alert colors. See setCriteria()}

\item{miss}{how missing data is treated. "last" if last value is repeated. 
It is currently the only option}
}
\value{
list containing the data, the alert indices, and the rules used to define the indices.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
# Parameters of the alert model
criteriaU = setCriteria(rule = "Af", val = c("tcrit"="22","preseas"="10","inccrit"="100"))
# Get, organize data 
cas = getCases(city = 3304557, cid10 = "A90", datasource=con) \%>\% 
     Rt(count = "casos",gtdist="normal", meangt=3, sdgt = 1) \%>\%
     mutate(inc = casos/pop*100000)
cli = getWU(stations = 'SBVT', vars=c("temp_min"), datasource=con) \%>\%
     mutate(temp_min = nafill(temp_min, rule = "arima")) 
# Calculate alert      
ale <- join_all(list(cas,cli),by="SE") 
res <- fouralert(ale, crit = criteriaU)
# Better visualization
tail(write.alerta(res))
}
