% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{fouralert}
\alias{fouralert}
\title{Define conditions to issue a four level alert Green-Yellow-Orange-Red.}
\usage{
fouralert(obj, crit, miss = "last", dy = 4)
}
\arguments{
\item{obj}{dataset with data to feed the alert, containing the variables specified in crit.}

\item{crit}{criteria for the alert colors. See setCriteria()}

\item{miss}{how missing data is treated. "last" if last value is repeated. 
It is currently the only option}

\item{dy}{if inc > 0, and rt was orange or red at least once in the past 
dy weeks -> level yellow. Default: dy=4}
}
\value{
returns an object of class "alerta" containing four elements: the data, 
the alert indices, and the rules used to define the indices.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario.
}
\examples{
# Parameters of the alert model
val = c(varcli ="temp_min", "clicrit"="22","limiar_preseason"="10","limiar_epidemico"="100")
criteria = setCriteria(rule="Af",values=val)
# Get, organize data 
cas = getCases(cities = 3304557, cid10 = "A90", datasource=con) \%>\% 
     Rt(count = "casos",gtdist="normal", meangt=3, sdgt = 1) \%>\%
     mutate(inc = casos/pop*100000)
cli = getWU(stations = 'SBGL', vars=c("temp_min"), datasource=con) \%>\%
     mutate(temp_min = nafill(temp_min, rule = "arima")) 
# Calculate alert      
ale <- plyr::join_all(list(cas,cli),by="SE") 
resf <- fouralert(ale, crit = criteria)
# Better visualization
tail(tabela.historico(resf),n=3)
}
