% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_functions.R
\name{GenTimeDist}
\alias{GenTimeDist}
\title{Calculates temperature dependent generation time distribution}
\usage{
GenTimeDist(serTemp, cid10 = "A90", GT.max = 10, smooth = "sinusoidal", nc = 1)
}
\arguments{
\item{serTemp}{: Temperature series}

\item{cid10}{"A90" for dengue (default), "A92": chikungunya, "A92.8": zika.}

\item{GT.max}{: maximum number of weeks to consider for generation time}

\item{smooth}{"sinusoidal" (default) or "raw"}
}
\value{
matrix with generation time distributions per week, one row per week, one column per SE.
}
\description{
Function to produce matrix of generation time distribution that depends on temperature.
For details, see the reference Codeco et al (2018).
}
\examples{
cli = getWU(stations = 'SBGL', vars=c("temp_med"), datasource=con) \%>\%
     mutate(temp_med = nafill(temp_med, rule = "arima")) 
gt <- GenTimeDist(cli$temp_med[1:50])
contour(gt, ylab="time", xlab="generation time (week)")
}
\references{
Codeco et al (2018) https://doi.org/10.1016/j.epidem.2018.05.011
}
