% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{geraMapa}
\alias{geraMapa}
\title{Plot the alert map for any state.}
\usage{
geraMapa(
  alerta,
  subset,
  se,
  cores = c("green", "yellow", "orange", "red"),
  legpos = "bottomright",
  titulo = "",
  filename,
  dir = "",
  shapefile,
  varid,
  varname,
  caption = TRUE,
  resol = 200
)
}
\arguments{
\item{alerta}{object created by update.alerta.}

\item{subset}{nomes das cidades que serão mostradas no mapa.}

\item{se}{epidemiological week (format = 201610).}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}

\item{legpos}{legend position. Default is bottomright. Check the other options in the function legend.}

\item{titulo}{title of the map}

\item{filename}{if present, the map is saved.}

\item{dir}{directory where map will be saved}

\item{shapefile}{shapefile containing polygons for the municipalities}

\item{varid}{shapefile variable indicating the geocode of the municipalities}

\item{varname}{name of the variable to be plotted}

\item{caption}{Default is TRUE}

\item{resol}{dpi png resolution, default is 200}
}
\value{
a map
}
\description{
Function to plot a map of the alert.
}
\examples{
# Parameters for the model
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo, cid10 = "A90", 
finalday= "2018-08-12",nowcasting="none")
geraMapa(alerta=res, subset = c(3300936, 3302403), se=201804, 
shapefile="shape/33MUE250GC_SIR.shp", varid="CD_GEOCMU", titulo="RJ-Norte")
}
