% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCases}
\alias{getCases}
\title{Get Case Data and aggregate per week and area}
\usage{
getCases(city, lastday = Sys.Date(), withdivision = TRUE,
  disease = "dengue", datasource = "data/sinan.rda")
}
\arguments{
\item{city}{city's geocode.}

\item{withdivision}{either FALSE if aggregation at the city level, or TRUE.}

\item{disease}{default is "dengue".}

\item{datasource}{data server or "data/sinan.rda" if using local test data.}

\item{finalday}{last day. Default is the last available.}
}
\value{
data.frame with the data aggregated per week according to disease onset date.
}
\description{
Create weekly time series from case data from server. The source is the SINAN.
}
\examples{
dC0 = getCases(city = c(330455), withdivision = TRUE) # Rio de Janeiro
head(dC0)
dC0 = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
tail(dC0)
dC0 = getCases(city = c(330455), lastday ="2014-03-10", withdivision = FALSE) # Rio de Janeiro
tail(dC0)
}

