% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCases}
\alias{getCases}
\title{Get Case Data and aggregate per week and area}
\usage{
getCases(city, lastday = Sys.Date(), disease = "dengue", datasource)
}
\arguments{
\item{city}{city's geocode.}

\item{disease}{default is "dengue".}

\item{datasource}{"db" if using the project database or "data/sinan.rda" if using local test data.}

\item{finalday}{last day. Default is the last available.}
}
\value{
data.frame with the data aggregated per week according to disease onset date.
}
\description{
Create weekly time series from case data from server. The source is the SINAN.
}
\examples{
dC0 = getCases(city = c(330455), lastday ="2014-03-10", datasource = "data/sinan.rda") 
dC0 = getCases(city = 4100301, datasource = con) 
head(dC0)
}

