% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCases}
\alias{getCases}
\title{Get Case Data and aggregate per week and area}
\usage{
getCases(city, withdivision = TRUE, disease = "dengue",
  datasource = "test")
}
\arguments{
\item{city}{city's geocode.}

\item{withdivision}{either FALSE if aggregation at the city level, or TRUE.}

\item{disease}{default is "dengue".}

\item{datasource}{data server or "test" if using local test data.}
}
\value{
data.frame with the data aggregated per week according to disease onset date.
}
\description{
Create weekly time series from case data from server. The source is the SINAN.
}
\examples{
res = getCases(city = c(330455), withdivision = TRUE) # Rio de Janeiro
head(res)
res = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
head(res)
}

