% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getCases}
\alias{getCases}
\title{Get Case Data and aggregate per week and area}
\usage{
getCases(city, lastday = Sys.Date(), cid10 = "A90", datasource)
}
\arguments{
\item{city}{city's geocode.}

\item{cid10}{cid 10 code. Dengue = "A90" (default), Chik = "A92.0", Zika = "A92.8",}

\item{datasource}{PostgreSQLConnection to project database .}

\item{finalday}{last day. Default is the last available.}
}
\value{
data.frame with the data aggregated per week according to disease onset date.
}
\description{
Create weekly time series from case data from server. The source is the SINAN.
}
\examples{
d <- getCases(city = 3302205, lastday ="2018-03-10", datasource = con) # dengue
d <- getCases(city = 3304557, cid10="A92.0", datasource = con) # chikungunya, until last day available
d <- getCases(city = c(3302205,230440), datasource = con) # dengue, two cities
tail(d)
}
