% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{getCidades}
\alias{getCidades}
\title{get list of cities.}
\usage{
getCidades(regional, uf, datasource)
}
\arguments{
\item{regional}{full name of the regional.}

\item{uf}{full name of the state.}

\item{datasource}{name of the database}
}
\value{
vector with names of the cities.
}
\description{
consult database to get list of cities
}
\examples{
getCidades(regional = "Metropolitana I", uf="Rio de Janeiro")
}

