% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getClima}
\alias{getClima}
\title{Legacy climate reader}
\usage{
getClima(
  cities,
  vars = c("temp_min", "temp_max", "temp_med", "umid_min", "ampT", "umid_med",
    "umid_max", "precip_tot", "precip_max"),
  finalday = Sys.Date(),
  iniSE = 201501,
  lastSE,
  datasource
)
}
\arguments{
\item{cities, vars, finalday, iniSE, lastSE, datasource}{Legacy arguments.}
}
\description{
Legacy climate reader
}
