% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{getdelaydata}
\alias{getdelaydata}
\title{Get delay data for analysis}
\usage{
getdelaydata(cities, years, cid10 = "A90", datasource)
}
\arguments{
\item{d}{dataset with case data containing at least three variables: the initial and final dates and a variable
identifying the epidemiological week (SEM_NOT).}

\item{tini}{variable indicating the initial date}

\item{tfim}{variable indicating the end date}

\item{SE}{variable indicating the epidemiological week}

\item{date.format}{date format. Default is "%d-%m-%Y"}

\item{truncdays}{Default is 183 days}

\item{plotar}{Default is TRUE}
}
\value{
list with d = data.frame with the epidemiological weeks; delay.tbl and delay.week 
are internal objects used for plotting.
}
\description{
Gets delay data for one or more cities. Internal function used in the delay fitting using inla.
}
\examples{
dados <- getdelaydata(cities=c(3302205, 3304557), datasource=con)  # Not run without connection
}
\author{
Claudia Codeco
}
