% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_incidence_functions.R
\name{getdelaydata}
\alias{getdelaydata}
\title{Get delay data for one or more cities for delay analysis}
\usage{
getdelaydata(cities, years = NULL, cid10 = "A90", lastday = Sys.Date(),
  datasource)
}
\arguments{
\item{cities}{vector with geocodes}

\item{years}{vector with set of years for analysis. Default (NULL) is to get all years of data available.}

\item{cid10}{disease code, Default is dengue. "A92.0" for chik, "A92.8" for zika}

\item{lastday}{last day of reporting (digitacao)}

\item{datasource}{valid connection to database}
}
\value{
list with d = data.frame with data.
}
\description{
Gets delay data for one or more cities. Internal function used in the delay fitting using inla.
}
\examples{
dados <- getdelaydata(cities=3304557, years=c(2016, 2017), cid10="A92", lastday = "2017-06-05" ,datasource=con)  # Not run without connection
}
