% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPop}
\alias{getPop}
\title{Get population data at municipal level.}
\usage{
getPop(cities, iniY = 2010, endY)
}
\arguments{
\item{cities}{list of geocodes.}

\item{iniY}{first year. Default is 2010}

\item{lastY}{last year. Default is the last available}
}
\value{
tibble with columns geocode, year, pop
}
\description{
Get population time series, using the brpop package. 
Currently has estimates from 2000 to 2021
}
\examples{
cities <- getCidades(uf = "Paraná")$municipio_geocodigo
pop <- getPop(cities, iniY = 2018, endY = 2019)
tail(pop)
}
