% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{getRegionais}
\alias{getRegionais}
\title{get list of regionais.}
\usage{
getRegionais(
  cities = NULL,
  uf,
  sortedby = "a",
  macroreg = FALSE,
  datasource = con,
  complete = FALSE
)
}
\arguments{
\item{cities}{cities' geocodes}

\item{uf}{full name of the state.}

\item{sortedby}{the options are: 'a' alphabetically, 'id' regional id number (only valid for regional), 
if available}

\item{macroreg}{TRUE if getRegionais should return macroreg instead of reg. Default: False}

\item{database}{name of the database}
}
\value{
vector with names of the regionais.
}
\description{
consult database to get list of regionais
}
\examples{
getRegionais(uf="Rio de Janeiro")
getRegionais(uf="Minas Gerais")
getRegionais(uf="Minas Gerais", macroreg = TRUE)
getRegionais(cities = c(3304128,3306107,3300159), uf="Rio de Janeiro")
getRegionais(cities = c(3304128,3306107,3300159), uf="Rio de Janeiro", macroreg = TRUE)
getRegionais(uf="Rio de Janeiro", sortedby = 'id')
}
