% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getTweet}
\alias{getTweet}
\title{Get Tweeter Data}
\usage{
getTweet(city, datasource = "test")
}
\arguments{
\item{city}{city's geocode.}

\item{datasource}{server or "test" if using test dataset.}
}
\value{
data.frame with weekly counts of people tweeting on dengue.
}
\description{
Create weekly time series from tweeter data from server. The
source of this data is the Observatorio da Dengue (UFMG).
}
\examples{
res = getTweet(city = c(330455)) # Rio de Janeiro
head(res)
}

