% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getTweet}
\alias{getTweet}
\title{Get Tweeter Data}
\usage{
getTweet(city, lastday = Sys.Date(), datasource = "data/tw.rda")
}
\arguments{
\item{city}{city's geocode.}

\item{datasource}{server or "data/tw.rda" if using test dataset.}

\item{finalday}{last day. Default is the last available.}
}
\value{
data.frame with weekly counts of people tweeting on dengue.
}
\description{
Create weekly time series from tweeter data from server. The
source of this data is the Observatorio da Dengue (UFMG).
}
\examples{
res = getTweet(city = c(330455)) # Rio de Janeiro
tail(res)
res = getTweet(city = c(330455), lastday = "2014-03-01") # Rio de Janeiro
tail(res)
}

