% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getTweet}
\alias{getTweet}
\title{Get Tweeter Data}
\usage{
getTweet(city, lastday = Sys.Date(), datasource)
}
\arguments{
\item{city}{city's geocode.}

\item{datasource}{server or "data/tw.rda" if using test dataset. 
Use the connection to the Postgresql server if using project data. See also DenguedbConnect
to open the database connection.}

\item{finalday}{last day. Default is the last available.}
}
\value{
data.frame with weekly counts of people tweeting on dengue.
}
\description{
Create weekly time series from tweeter data from server. The 
source of this data is the Observatorio da Dengue (UFMG).
}
\examples{
res = getTweet(city = c(330455), lastday = "2014-03-01", datasource = con)
res = getTweet(city = c(330455), datasource = "data/tw.rda") 
tail(res)
}

