% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getWU}
\alias{getWU}
\title{Get Climate Data}
\usage{
getWU(stations, var = "all", city = c(), finalday = Sys.Date(),
  datasource = "data/WUdata.rda")
}
\arguments{
\item{stations}{vector with the stations codes (4 digits).}

\item{var}{climate variables (default var="all": all variables available )}

\item{city}{vector with the cities codes (6 digits). Not implemented yet.}

\item{finalday}{last day. Default is the last available. Format = Year-month-day.}

\item{datasource}{use "db" to refer to the sql server. Default is the test dataset.}
}
\value{
data.frame with the weekly data (cidade estacao data tmin tmed tmax umin umed umax pressaomin pressaomed pressaomax)
}
\description{
Create weekly time series from meteorological station data in server
}
\examples{
res = getWU(stations = c('SBRJ','SBJR','SBAF','SBGL'))
res = getWU(stations = 'SBRJ', var="tmin", finalday = "2014-10-10")
}

