% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getWU}
\alias{getWU}
\title{Get Climate Data}
\usage{
getWU(stations, vars = "temp_min", finalday = Sys.Date(), datasource)
}
\arguments{
\item{stations}{station code (4 digits).}

\item{vars}{climate variables (default var="all": all variables available )}

\item{finalday}{last day. Default is the last available. Format = Y-m-d.}

\item{datasource}{Use the connection to the Postgresql server if using project data.}
}
\value{
data.frame with the weekly data
}
\description{
Create weekly time series from meteorological station data in server
}
\examples{
res = getWU(station = 'SBRJ', vars="temp_min", finalday = "2014-10-10", 
datasource= con)
tail(res)
}

