% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getWU}
\alias{getWU}
\title{Get Climate Data}
\usage{
getWU(
  stations,
  vars = "temp_min",
  finalday = Sys.Date(),
  iniSE = 201001,
  datasource = con
)
}
\arguments{
\item{stations}{station code (4 digits).}

\item{vars}{vector with meteorological variables. Available variables: "temp_min" (default), 
"temp_max","temp_med","data_dia","umid_min","umid_med","umid_max","pressao_min","pressao_med","pressao_max"}

\item{finalday}{last day. Default is the last available. Format = Y-m-d.}

\item{datasource}{Use "data/WUdata.rda" to use test dataset. Use the connection to the Postgresql server if using project data. See also DenguedbConnect
to open the database connection.}
}
\value{
data.frame with the weekly data (cidade estacao data temp_min tmed tmax umin umed umax pressaomin pressaomed pressaomax)
}
\description{
Create weekly time series from meteorological station data in server taking the mean of the daily values.
}
\examples{
NOT USE: con <- dbConnect(RSQLite::SQLite(), "../../AlertaDengueAnalise/mydengue.sqlite")
res = getWU(stations = c('SBRJ','SBGL'), vars="temp_min", iniSE = 201201)
res = getWU(stations = 'SBRJ', vars=c("temp_min", "temp_med"))
tail(res)
}
