% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{getWUstation}
\alias{getWUstation}
\title{get meteorological stations}
\usage{
getWUstation(cities, datasource = con)
}
\arguments{
\item{cities}{vector with geocodes}

\item{datasource}{connection to the project database}
}
\value{
data.frame
}
\description{
Get the meteorological stations associated with one or more cities
}
\examples{
getWUstation(cities = 3304557)
cidades <- getCidades(regional = "Sete Lagoas", uf = "Minas Gerais")
getWUstation(cities = cidades$municipio_geocodigo)
}
