% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{insertCityinAlerta}
\alias{insertCityinAlerta}
\title{Initial setup of a new city in the alerta system.}
\usage{
insertCityinAlerta(city, id_regional, regional, senha)
}
\arguments{
\item{city}{geocode of the city. Mandatory}

\item{id_regional}{numerical id of the 'Regional da saude'}

\item{regional}{name of the 'Regional da saude'}

\item{estacao_wu_sec}{main weather station}
}
\value{
to be defined
}
\description{
Initial setup of a new city in the alerta system.  Can be integrated later with 
the delay model and write.parameters
}
\examples{
insertCityinAlerta(city=3200300, id_regional=0, regional = "ES-MN-AlfredoChaves")
}
