% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{isOrange}
\alias{isOrange}
\title{Raise orange alert if sustained transmission is detected.}
\usage{
isOrange(obj, pvalue = 0.9, lag = 3)
}
\arguments{
\item{obj}{object created by the Rt function.}

\item{pvalue}{probability of wrongly rejecting the hypothesis Rt > 1}

\item{lag}{count the number of weeks within the last lag weeks with conditions = TRUE}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE
}
\description{
"Orange" is defined by effective reproductive number greater than 1. 
Rt must be computed before using this function.
}
\examples{
cas = getCases(city = c(330455), withdivision = FALSE, datasource="data/sinan.rda")
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
ora = isOrange(obj = casr, pvalue = 0.9, lag= 3) 
head(ora)
x = 1:length(ora$SE)
plot(x, ora$Rt, type="l", xlab= "weeks", ylab = "Rt")
lines(x, ora$upr, lty = 3)
lines(x, ora$lwr, lty = 3)
abline(h = 1, col =2)
points(x[ora$oweek==1], ora$Rt[ora$oweek==1], col="orange", pch=16)
}

