% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{lastDBdate}
\alias{lastDBdate}
\title{Returns the most recent date present in the database table.}
\usage{
lastDBdate(tab, city = NULL, station = NULL, datasource)
}
\arguments{
\item{tab}{table in the database. Either (sinan, clima_wu, tweet ou historico).}

\item{city}{city geocode, if empty, whole dataset is considered. Not implemented yet.}

\item{station}{wu station.}

\item{datasource}{}
}
\value{
most recent date
}
\description{
Useful to check if the database is up-to-date.
}
\examples{
lastDBdate(tab="tweet",datasource=con)
lastDBdate(tab="tweet", city=330240,datasource=con)
lastDBdate(tab="sinan", city=330240,datasource=con)
lastDBdate(tab="clima_wu", station="SBAF",datasource=con)  
}

