% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{nafill}
\alias{nafill}
\title{methods to substitute NAs.Use the function na.approx from package zoo.}
\usage{
nafill(v, rule, maxgap = 4)
}
\arguments{
\item{v}{vector with missing elements.}

\item{rule}{rule for filling the missing cells. "zero" just fills them with 0; "linearinterp"
interpolate using tzoo::na.approx. In this case, the tails are not filled.}

\item{maxgap}{maximum number of consecutive NAs to fill. Longer gaps will be left i=unchanged}
}
\value{
vector with replaced NA.
}
\description{
collection of imputation procedures
}
\examples{
v <- c(1,2,3,NA,5,6,NA,NA,9,10,NA,NA)
nafill(v, rule = "zero")
nafill(v, rule = "linear")
}

