% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{plot.alerta}
\alias{plot.alerta}
\title{Plot the time series of warnings.}
\usage{
\method{plot}{alerta}(obj, var, cores = c("#0D6B0D", "#C8D20F", "orange",
  "red"), ini = 201001, fim = 202001, ylab = var, yrange)
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{var}{to be ploted in the graph, usually cases when available.}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}
}
\value{
a plot
}
\description{
Function to plot the output of
}
\examples{
 # See fouralert function
}

