% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{plot.alerta}
\alias{plot.alerta}
\title{Plot the time series of warnings.}
\usage{
\method{plot}{alerta}(obj, geocodigo, var = "casos",
  cores = c("#0D6B0D", "#C8D20F", "orange", "red"), ini = 201001,
  fim = 202001, ylab = var, yrange, salvar = FALSE,
  nome.fig = "grafico", datasource = con)
}
\arguments{
\item{obj}{object created by tabela_historico()}

\item{geocodigo}{city's geocode.}

\item{var}{variable to be plotted, usually "cases", or "p_inc100k".}

\item{cores}{colors corresponding to the levels 1, 2, 3, 4.}

\item{ini}{first epidemiological week. If not stated, use the first one available}

\item{fim}{last epidemiological week. If not stated, use the last one available}

\item{ylab}{y axis label}

\item{yrange}{y axis range}

\item{salvar}{TRUE to save the figure (default is FALSE)}

\item{nome.fig}{figure name (default is "grafico")}
}
\value{
a plot
}
\description{
Function to plot the output of tabela_historico.
}
\examples{
# Parameters for the model
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo, cid10 = "A90", 
finalday= "2018-08-12",nowcasting="none")
restab <- tabela_historico(res)
plot.alerta(restab, geocodigo = cidades$municipio_geocodigo, var="casos")
}
