% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{read.cases}
\alias{read.cases}
\title{Function to extract case count for covered municipalities}
\usage{
read.cases(start_year, end_year, con, mun_list = NULL)
}
\description{
Function \code{read.cases} extract notification data from database and aggregate by epiweek. Used by mem.
}
\author{
Marcelo F C Gomes
}
\keyword{internal}
