% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{read.cases}
\alias{read.cases}
\title{Function to extract case count for covered municipalities}
\usage{
read.cases(start_year, end_year, datasource = con, mun_list)
}
\arguments{
\item{start_year}{first year of the time series}

\item{end_year}{last year of the time series}

\item{datasource}{Infodengue connection}

\item{mun_list}{vector with the municipalities' 7 digit geocodes}
}
\description{
Function \code{read.cases} extract notification data from database 
and aggregate by epiweek. Used by mem.
}
\examples{
dd <- read.cases(2019, 2020, mun_list = c(4108304, 3300936))
}
\author{
Marcelo F C Gomes
}
