% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{read.parameters}
\alias{read.parameters}
\title{Get city alert parameters.}
\usage{
read.parameters(city, region, state, datasource)
}
\arguments{
\item{city}{geocode}

\item{region}{regional's name}

\item{state}{state's name}

\item{datasource}{SQL connection to the database}
}
\value{
dataframe with all parameters
}
\description{
Read the alert parameters for a set of  cities from the database, to be used in the update.alert. 
Currently, the parameters are:"codigo_estacao_wu", "limiar_preseason", "limiar_posseason",
"limiar_epidemico, "estacao_wu_sec".
}
\examples{
read.parameters(city = 3118601, datasource = con)
read.parameters(region="Norte", state = "Rio de Janeiro", datasource = con)
read.parameters(state = "Rio de Janeiro", datasource = con)
read.parameters(datasource = con) #set no filter to get all cities
}
