% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_functions.R
\name{Rt}
\alias{Rt}
\title{Computes the effective reproductive number using alternative 
distributions for the generation interval.}
\usage{
Rt(obj, count = "casos", gtdist, meangt, sdgt, CI = "beta", alpha = 0.95,
  a0 = 2, b0 = 3)
}
\arguments{
\item{meangt}{if gtdist = "delta" it is the exact period between primary and 
secondary infections). If gtdist = "normal", it is the mean generation time.}

\item{sdgt}{if gtdist = "normal", it is the standard deviation of the generation time 
distribution.}

\item{CI}{Model used to compute the confidence interval. Possible choice: "beta".}

\item{object}{with the format of getcases. A data.frame with variables "casos" and "SE".}
}
\value{
data.frame with estimated Rt and confidence intervals.
}
\description{
Calculates the effective reproductive number from the growth rate 
of cases. Uses formula 4.2 in Wallinga and Lispitch (2007). Confidence interval assume
ratio between two Poissons (see Luis Max documentation).
}
\examples{
res = getCases(city = c(330455), withdivision = FALSE, datasource="data/sinan.rda") # Rio de Janeiro
# Rt original
rt<-Rtoriginal(obj = res, count = "casos", meangt=3)
plot(rt$Rt, type="l", xlab = "weeks", ylab = "Rt")
lines(rt$lwr,lty=3)
lines(rt$upr,lty=3)
abline(h = 1, col = 2)
# Rt delta and normal
rtdelta<-Rt(obj = res, count = "casos", gtdist="delta", meangt=3)
rtnorm<-Rt(obj = res, count = "casos", gtdist="normal", meangt=3, sdgt = 1)
lines(rtdelta$Rt, col = 3)
lines(rtdelta$lwr,lty = 3, col = 3)
lines(rtdelta$upr,lty = 3, col = 3)
lines(rtnorm$Rt, col = 4)
lines(rtnorm$lwr,lty = 3, col = 4)
lines(rtnorm$upr,lty = 3, col = 4)
legend(30,3,c("original","delta","normal"),lty=1, col = c(1,3,4), cex = 0.7)
}

