% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_functions.R
\name{Rt}
\alias{Rt}
\title{Computes the effective reproductive number using alternative 
distributions for the generation interval.}
\usage{
Rt(
  obj,
  count = "casos",
  group = "cidade",
  gtdist,
  meangt,
  sdgt,
  CI = "beta",
  alpha = 0.95,
  a0 = 2,
  b0 = 3
)
}
\arguments{
\item{obj}{a data.frame with variables "casos" and "SE", ideally from getCases().}

\item{group}{a variable in data.frame indicating a stratification (city, p.e.)}

\item{meangt}{if gtdist = "delta" it is the exact period between primary and 
secondary infections). If gtdist = "normal", it is the mean generation time.}

\item{sdgt}{if gtdist = "normal", it is the standard deviation of the generation time 
distribution.}

\item{CI}{Model used to compute the confidence interval. Possible choice: "beta".}
}
\value{
data.frame with estimated Rt and confidence intervals.
}
\description{
Calculates the effective reproductive number from the growth rate 
of cases. Uses formula 4.2 in Wallinga and Lispitch (2007). Confidence interval assume
ratio between two Poissons (see Luis Max documentation).
}
\examples{
d <- getCases(cities = c(3302205, 3304557), lastday ="2018-03-10") # dengue
Rt delta and normal
rtdelta <- Rt(obj = d, count = "casos", gtdist="delta", meangt=3)
rtnorm <- Rt(obj = d, count = "casos", gtdist="normal", meangt=3, sdgt = 1)
d <- d \%>\% Rt(count = "casos", gtdist="delta", group = "cidade", meangt=3)
}
