% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_functions.R
\name{Rtoriginal}
\alias{Rtoriginal}
\title{Computes the effective reproductive number using a Alerta's original formulation.}
\usage{
Rtoriginal(obj, count = "casos", meangt, CI = "beta", alpha = 0.95,
  a0 = 2, b0 = 3)
}
\arguments{
\item{obj}{object with the format of getcases. A data.frame with variables "casos" and "SE".}

\item{count}{case counts (adjusted or nonadjusted incidence).}

\item{meangt}{the exact period between primary and secondary infections}

\item{CI}{Model used to compute the confidence interval. "beta" is the only choice.}
}
\value{
data.frame with estimated Rt and confidence intervals
}
\description{
Calculates the effective reproductive number from the growth rate 
of cases. The The original formulation is simply the ratio between 3-weeks accumulated
cases at week (t+1) and the 3-weeks accumulated cases at week t.
}
\examples{
res = getCases(city = c(330455), withdivision = FALSE) # Rio de Janeiro
# Rt original
rt<-Rtoriginal(obj = res, count = "casos", meangt=3)
plot(rt$Rt, type="l", xlab = "weeks", ylab = "Rt")
lines(rt$lwr,lty=3)
lines(rt$upr,lty=3)
abline(h = 1, col = 2)
}

