% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{SE2date}
\alias{SE2date}
\title{Return the first day of the Epidemiological Week}
\usage{
SE2date(se)
}
\arguments{
\item{SE}{string vector with dates to be converted, format 201420}
}
\value{
data.frame with SE and first day. Only works for years >= 2010.
}
\description{
Return to which epidemiological week belongs a given day (only >=2010)
}
\examples{
SE2date(201512)
SE2date(se = c(201401:201409))
}

