% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{seqSE}
\alias{seqSE}
\title{Sequence of epidemiological weeks}
\usage{
seqSE(from, to)
}
\arguments{
\item{from}{first week in format 201401}

\item{to}{first week in format 201401}
}
\value{
data.frame with the epidemiological weeks and corresponding extreme days.
}
\description{
Creates a sequence of epidemiological weeks and respective end and final days
}
\examples{
seqSE(201502, 201510)
}

