% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{setCriteria}
\alias{setCriteria}
\title{Define rules to issue a four level alert Green-Yellow-Orange-Red.}
\usage{
setCriteria(rule = NULL, values = NULL, delays = list(delayy = c(0,
  0), delayo = c(0, 0), delayr = c(0, 0)))
}
\arguments{
\item{rule}{either a built-in rule ("Af" or "Aw") or a list with three elements defining criteria for transition to yellow (level 2),
orange (level 3) and red (level 4). See description.}

\item{values}{named vector of values for the critical parameters. Use character.}

\item{delays}{list with three elements, each one is a vector: c(delay to turn on, delay to turn off)}
}
\value{
list with rules. To be useful, this list must contain variables that match
those in the data.
}
\description{
The criteria for transition between colors (alert levels) can be 
chosen from existing rules or can be specified by the user. The built in rules are: 
Af (minimum temperature defines yellow) and Aw (humidity does).
}
