% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{tabela_historico}
\alias{tabela_historico}
\title{Write the alert object into the database.}
\usage{
tabela_historico(obj, versao = Sys.Date())
}
\arguments{
\item{obj}{object created by the pipeline.}

\item{versao}{Default is current's date}

\item{ini_se}{first week of the table. Default is the first date in obj.To do.}

\item{last_se}{last week of the table. Default is the last date in obj. To do.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo, cid10 = "A90", 
finalday= "2013-01-10")
restab <- tabela_historico(res)
}
