% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{tabela_historico}
\alias{tabela_historico}
\title{Convert the alert object into a data.frame and calculate indicators}
\usage{
tabela_historico(obj, iniSE, lastSE, type = "notified", versao = Sys.Date())
}
\arguments{
\item{obj}{object created by the pipeline.}

\item{type}{"notified", if it should return total counts, disregarding the 
final classification. For compatibility reasons.}

\item{versao}{Default is current's date}

\item{ini_se}{first week of the table. Default is the first date in obj.}

\item{last_se}{last week of the table. Default is the last date in obj. To do.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to organize the alert results for easy reading and inserting 
in the database. Also computes receptivity, transmission and incidence levels.
}
\examples{
# Several cities at once:
cidades <- getCidades(uf = "Mato Grosso", datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo[1:3], cid10 = "A90", 
finalday= "2018-01-10")
restab <- tabela_historico(res, iniSE = 201701,type = "all") 
tail(restab)
# One city:
res <- pipe_infodengue(cities = 3304557, cid10 = "A90", 
finalday= "2015-01-10")
restab <- tabela_historico(res) 
tail(restab)
}
