% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{tabela_historico_intra}
\alias{tabela_historico_intra}
\title{Convert the alert object into a data.frame and calculate indicators for cities with subdivisions}
\usage{
tabela_historico_intra(obj, iniSE, lastSE, versao = Sys.Date())
}
\arguments{
\item{obj}{object created by the pipeline.}

\item{versao}{Default is current's date}

\item{ini_se}{first week of the table. Default is the first date in obj.}

\item{last_se}{last week of the table. Default is the last date in obj. To do.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to organize the alert results for easy reading and inserting 
in the database. Specific for municipalities with subdivisions.
}
\examples{
NOT RUN without connection
# Rio de Janeiro
alerio <- pipe_infodengue_intra(city = 3304557, se=202105, delaymethod="bayesian",
cid10 = "A90", dataini = "sinpri")
restab <- tabela_historico_intra(alerio, iniSE = 201801) 
tail(restab)
}
