% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{temp.predict}
\alias{temp.predict}
\title{Fit arima to fill in missing data at the end of temperature time series.}
\usage{
temp.predict(v, plotar = FALSE)
}
\arguments{
\item{v}{vector with temperature data.}
}
\value{
vector with replaced NA.
}
\description{
function for extrapolating temperature using arima
}
\examples{
tail(cliSBCB)
temp.predict(v=cliSBCB[,3], plotar = T)
}

