% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{twoalert}
\alias{twoalert}
\title{Define conditions to issue a 2 level alert Green/Yellow.}
\usage{
twoalert(obj, cy)
}
\arguments{
\item{obj}{dataset from the mergedata function containing at least SE, and the 
variables used for alert calculation.}

\item{cy}{criteria to set the yellow alarm, written as a vector with three elements.
The first is the condition (see the example), the second is the number of times the
condition must be tru to issue the yellow alert, and the third, the number of weeks
false to turn off the alert (green).}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
This function is meant to be used when case data is absent. 
In this scenario, only two levels exist: Yellow if environmental conditions 
required for positive mosquito population growth are detected, or if social activity
increases. Green otherwise. But clearly, the user can define any rule.
}
\examples{
tw = getTweet(city = c(330455), datasource = "data/tw.rda") 
cli = getWU(stations = 'SBRJ', datasource="data/WUdata.rda")
d<- mergedata(tweet = tw, climate = cli)
crity <- c("temp_min > 22", 3, 3)
alerta <- twoalert(d, cy = crity)
head(alerta$indices)
plot.alerta(alerta, var="temp_min")
}

