% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{update.alerta}
\alias{update.alerta}
\title{Define conditions to issue a 4 level alert Green-Yellow-Orange-Red for any city or region.}
\usage{
\method{update}{alerta}(city, region, state, pars, crit, cid10 = "A90",
  writedb = FALSE, datasource, sefinal, adjustdelay = T)
}
\arguments{
\item{city}{city's geocode (6 or 7 digits).}

\item{region}{full name of 'regional' or state (same name present in the database).}

\item{state}{full name of state (same name present in the database). Required if there are more than one region with the same name.}

\item{pars}{list of parameters for the alerta, defined in config.R}

\item{crit}{criteria for the alert colors, defined in configglobal.R}

\item{cid10}{default is A90 (dengue). Chik = A920, Zika = A92.8}

\item{writedb}{TRUE if it should write into the database, default is FALSE.}

\item{sefinal}{if given, it stops at that week}

\item{adjustdelay}{Default is TRUE, if F, there is no delay adjustment and estimated = observed.}

\item{temp_station}{code of the meteorological station for temperature. 
If not provided, use default from database. To be implemented.}
}
\value{
data.frame with the week condition and the number of weeks within the 
last lag weeks with conditions = TRUE.
}
\description{
Yellow is raised when environmental conditions required for
positive mosquito population growth are detected, green otherwise.Orange 
indicates evidence of sustained transmission, red indicates evidence of 
an epidemic scenario. For most places weather stations are provided in the database and choice
is authomatic according to data quality. But it can also be provided manually (not implemented).
}
\examples{
# Parameters for the model
criteria = list(crity = c("temp_min > tcrit", 3, 1),
crito = c("p1 > 0.95 & inc > preseas & temp_min >= tcrit", 3, 1),
critr = c("inc > inccrit", 2, 2))
gtdist="normal"; meangt=3; sdgt = 1.2
pars.RJ <- NULL
pars.RJ[["Norte"]] <- list(pdig = c(2.997765,0.7859499),tcrit=22, inccrit = 100, preseas=8.28374162389761, posseas = 7.67878514885295, legpos="bottomright")
# Running the model:
res <- update.alerta(city = 3205309, pars = pars.ES[["Central"]], crit = criteria, datasource = con)
res <- update.alerta(region = "Metropolitana I", pars = pars.RJ, crit = criteria, datasource = con,sefinal=201613)
tail(res$data)
}
