% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alerta}
\alias{write.alerta}
\title{Write the alert into the database.}
\usage{
write.alerta(obj, write = "no", version = Sys.Date())
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database. It only writes one city at a time. It is recommended that the end data is specified.
If this is the first time a city is included in the dataset, than use newcity = TRUE. This will force to insert from the beginning. Or if you want to update,
you can define the ini-end dates or define the last n weeks.
}
\examples{
cli = getWU(stations = 'SBCB', datasource=con)
cas = getCases(city = 330020, withdivision = FALSE, datasource=con)
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = cli)
crity <- c("temp_min > 22", 3, 3)
crito <- c("p1 > 0.9", 3, 3)
critr <- c("inc > 100", 3, 3)
alerta <- fouralert(d, cy = crity, co = crito, cr = critr, pop=1000000)
res <- write.alerta(alerta, write="no")
tail(res)
}

