% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alerta}
\alias{write.alerta}
\title{Write the alert into the database.}
\usage{
write.alerta(obj, ini, end, write = "no", version = Sys.Date(),
  NAaction = "omitcasesmedNA")
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{ini}{first epidemiological week of the period to be written.}

\item{end}{last epidemiological week of the period to be written.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
tw = getTweet(city = c(330455), datasource = "data/tw.rda") 
cli = getWU(stations = 'SBRJ', datasource="data/WUdata.rda")
cas = getCases(city = c(330455), withdivision = FALSE, datasource="data/sinan.rda")
casfit<-adjustIncidence(obj=cas)
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = clima)
crity <- c("temp_min > 22", 3, 3)
crito <- c("p1 > 0.9", 3, 3)
critr <- c("inc > 100", 3, 3)
alerta <- fouralert(d, cy = crity, co = crito, cr = critr, pop=1000000)
res <- write.alerta(alerta)
tail(res)
}

