% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alerta}
\alias{write.alerta}
\title{Write the alert object into the database.}
\usage{
write.alerta(obj, write = "no", version = Sys.Date())
}
\arguments{
\item{obj}{object created by the twoalert and fouralert functions.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database. It only writes one city at a time. It is recommended that the end data is specified.
If this is the first time a city is included in the dataset, than use newcity = TRUE. This will force to insert from the beginning. Or if you want to update,
you can define the ini-end dates or define the last n weeks.
}
\examples{
# Getting the data (requires a con connection)
tw = getTweet(city = 2304400, datasource = con) 
cli = getWU(stations = 'SBFZ', vars=c("temp_min", "umid_min"), datasource=con)
cas = getCases(city = 2304400, cid10="A92.0",datasource=con)
# Organizing the data
casfit<-adjustIncidence(obj=cas, method="fixedprob")
casr<-Rt(obj = casfit, count = "tcasesmed", gtdist="normal", meangt=3, sdgt = 1)
d<- mergedata(cases = casr, tweet = tw, climate = cli)
# Parameters of the alert model (usually set up in the globalconfig and config files)
criteriaU = list(crity = c("temp_min > tcrit & inc > 0", 3, 1),
crito = c("p1 > 0.95 & inc > preseas", 3, 1),
critr = c("inc > inccrit", 2, 2))
gtdist="normal"; meangt=3; sdgt = 1.2
pars.RJ <- NULL
pars.RJ[["Norte"]] <- list(pdig = c(2.997765,0.7859499),tcrit=22, ucrit=87, inccrit = 100, preseas=8.28374162389761, 
posseas = 7.67878514885295, legpos="bottomright")
# Running the alert
ale <- fouralert(d, pars = pars.RJ[["Norte"]], crit = criteriaU, pop = 1000000)
ale <- fouralert(d, pars = pars.ES[["Central"]], crit = criteriaU, pop = 1000000)
# For a more useful output
res <- write.alerta(ale, write="no")
tail(res)
}
