% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write.alerta}
\alias{write.alerta}
\title{Write historico.alerta into the database.}
\usage{
write.alerta(d, datasource = con)
}
\arguments{
\item{d}{object created by historico.alerta()}

\item{datasource}{posgreSQL conn to project's database}
}
\value{
the same data.frame from the input
}
\description{
Function to write the pipeline results into the database. Receives the object created by the function historico.alerta
}
\examples{
# Parameters for the model
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo[1], cid10 = "A90", 
finalday= "2018-08-12",nowcasting="none")
restab <- tabela_historico(res[1]) 
write.alerta(restab[1:5,])
}
