% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write_alerta}
\alias{write_alerta}
\title{Write historico_alerta into the database.}
\usage{
write_alerta(d, datasource = con)
}
\arguments{
\item{d}{object created by tabela_historico()}

\item{datasource}{posgreSQL conn to project's database}
}
\value{
the same data.frame from the input
}
\description{
Function to write the pipeline results into the database. 
Receives the object created by the function historico.alerta.
}
\examples{
# Parameters for the model
cidades <- getCidades(regional = "Norte",uf = "Rio de Janeiro",datasource = con)
res <- pipe_infodengue(cities = cidades$municipio_geocodigo[1], cid10 = "A90", 
finalday= "2016-08-12",nowcasting="none")
restab <- tabela_historico(res) 
write_alerta(restab)
}
