% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write_alertaRio}
\alias{write_alertaRio}
\title{Write the Rio de janeiro alert into the database.}
\usage{
write_alertaRio(obj, datasource = con)
}
\arguments{
\item{obj}{object created by the tabela_historico_intra function and contains 
alerts for each locality.}

\item{datasource}{connection to infodengue database}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
alerio <- pipe_infodengue_intra(city = 3304557, datarelatorio=202105, 
delaymethod="bayesian", cid10 = "A90", dataini = "sinpri")
restab <- tabela_historico_intra(alerio, iniSE = 201801) 
write_alertaRio(restab[1:10,])
}
