% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_functions.R
\name{write_alertaRio}
\alias{write_alertaRio}
\title{Write the Rio de janeiro alert into the database.}
\usage{
write_alertaRio(obj, write = "no", datasource = con, version = Sys.Date())
}
\arguments{
\item{obj}{object created by the alertRio function and contains alerts for each APS.}

\item{write}{use "db" if data.frame should be inserted into the project database,
or "no" (default) if nothing is saved.}

\item{version}{default is the current date}
}
\value{
data.frame with the data to be written.
}
\description{
Function to write the alert results into the database.
}
\examples{
params <- c(varcli ="temp_min", clicrit=22, limiar_epidemico=100, limiar_preseason = 14.15)
criter <- setCriteria(rule="Af", values = params)
alerio2 <- alertaRio(naps = c(1,2), crit = criter, datasource=con, se = 201201)
res <- write_alertaRio(alerio2, write="no")
tail(res)  
}
