% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{write_parameters}
\alias{write_parameters}
\title{City's parameterization.}
\usage{
write_parameters(city, cid10, params, overwrite = FALSE, datasource = con)
}
\arguments{
\item{params}{vector of the names of the params to be inserted in the table. Limiar is given as incidence. 
It can be a subset of the default.}
}
\value{
the new line in the parameters table
}
\description{
Write the alert parameters for each city into the database, to be used in the update.alert. 
Currently, the parameters are: "limiar_preseason", "limiar_posseason",
"limiar_epidemico,"varcli", "varcli2", "clicrit", "clicrit2" , "cid10", "codmodelo". 
City must be already in the regionais table.
}
\examples{
params = data.frame(municipio_geocodigo = 3506003,limiar_preseason = 4.50243, limiar_posseason = 3.962566, 
limiar_epidemico = 67.72364, varcli = "temp_min", clicrit = 22, cid10 = "A90", codmodelo = "Af") 
res = write_parameters(newpars)
}
